% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_structure.R
\name{admixturePlots}
\alias{admixturePlots}
\title{Generate a pdf with admixture plots for each K

Takes as input the output files from strucutre runs and generates a pdf with plots of admixture}
\usage{
admixturePlots(
  x,
  xdir = NULL,
  labels = NULL,
  method = "structure",
  save.as = file.path(getwd(), "admixturePlots.pdf"),
  userun = 1
)
}
\arguments{
\item{x}{Character vector with input files containing q-matrices. For method='structure', the files must be those ending in 'log_f'.}

\item{xdir}{Optional path to the directory containing all STRUCTURE output files plus the '.*_sampleIDs.txt' file generated by run_structure. Default NULL. If supplied, this overides save.as, method, labels, and x arguments.}

\item{labels}{Character vector with names of individuals. Default NULL.}

\item{method}{Character string. For now, must be "structure".}

\item{save.as}{Character string with path/name to use for the output PDF file. Default is to save the output in the current directory with the name "admixturePlots.pdf".}

\item{userun}{Number or numerical vector indicating which runs should be used for admixture plots. Default is 1 (the first run). When multiple runs are used, the mean is used across runs after aligning clusters.}
}
\value{
NULL; generates pdf with a barplot of admixture for each K
}
\description{
Generate a pdf with admixture plots for each K

Takes as input the output files from strucutre runs and generates a pdf with plots of admixture
}
