% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.outer.R
\name{bbox2points}
\alias{bbox2points}
\title{corner coordinates of area described by bounding box

Returns a 2x4 matrix with coordinates x,y of the four corner points for a region with an extent determined by the input bounding box.}
\usage{
bbox2points(bb)
}
\arguments{
\item{bb}{2x2 numerical matrix (bounding box matrix) with x,y rows and and min, max columns}
}
\value{
2x4 matrix with coordinates x y of points at the corners the region with extent defined by 'bb'
}
\description{
corner coordinates of area described by bounding box

Returns a 2x4 matrix with coordinates x,y of the four corner points for a region with an extent determined by the input bounding box.
}
