% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.if.executable.R
\name{check.if.executable}
\alias{check.if.executable}
\title{Check Executable

Test if a file is executable}
\usage{
check.if.executable(exe.path)
}
\arguments{
\item{exe.path}{Character string with full path to the file to check.}
}
\value{
Integer 0 if exe.path is executable, otherwise 1.
}
\description{
Check Executable

Test if a file is executable
}
