% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{check_mcmcpath_contents}
\alias{check_mcmcpath_contents}
\title{Check MCMC Path Contents

Checks that all of the EEMs output files are present in the output directory}
\usage{
check_mcmcpath_contents(mcmcpath)
}
\arguments{
\item{mcmcpath}{Path the directory containing EEMs output files for a particular chain}
}
\value{
NULL if all expected files generated by EEMs are present
}
\description{
Check MCMC Path Contents

Checks that all of the EEMs output files are present in the output directory
}
