% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{check_plot_params}
\alias{check_plot_params}
\title{Check Plot Parameters

Internal function used for making EEMS plots}
\usage{
check_plot_params(pars)
}
\arguments{
\item{pars}{List of paramaters}
}
\value{
List of parameters for plotting
}
\description{
Check Plot Parameters

Internal function used for making EEMS plots
}
