% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_miscwrappers.R
\name{config_miscwrappers}
\alias{config_miscwrappers}
\title{config_miscwrappers function

This function is used to tell misc.wrappers where a program (executable file) is located. You only need to run this function once for a particular program.}
\usage{
config_miscwrappers(exe.paths = NULL)
}
\arguments{
\item{exe.paths}{Character string or vector with the full path(s) to one or more executable files for programs that misc.wrappers uses.}
}
\value{
If exe.paths is non-NULL, the function will update settings and return 0 if successful. If exe.paths is NULL, the function returns a dataframe with all of the path settings.
}
\description{
config_miscwrappers function

This function is used to tell misc.wrappers where a program (executable file) is located. You only need to run this function once for a particular program.
}
