% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords2sp.poly.R
\name{coords2sp.poly}
\alias{coords2sp.poly}
\title{coords2sp.poly function

Convert a coordinates matrix into a SpatialPolygons object of circles centered at input coordinates}
\usage{
coords2sp.poly(coords.mat, r = 0.01)
}
\arguments{
\item{coords.mat}{Numerical matrix of coordinates with longitude and latitude columns, respectively; coordinates must be in decimal degrees format.}

\item{r}{Number indicating radius of circles, in degrees.}
}
\value{
SpatialPolygons object (see sp package)
}
\description{
coords2sp.poly function

Convert a coordinates matrix into a SpatialPolygons object of circles centered at input coordinates
}
