% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{cpSNPsVCF}
\alias{cpSNPsVCF}
\title{copy STACKS SNPs VCF files using informative names

Copies each of the 'populations.snps.vcf' files generated by the 'populations' program of the STACKS pipeline, and uses a more informative filename. This function will only work if the 'populations_setup' function was used for running populations.}
\usage{
cpSNPsVCF(indir, outdir, groupname = NULL)
}
\arguments{
\item{indir}{Character string with the parent directory where STACKS output files and directories are held for a particular group.}

\item{outdir}{Character string with path to directory where 'populations.snps.vcf' files should be copied. These will be renamed with the groupname stacks parameters.}

\item{groupname}{Character string with name of the group (usually a taxon). Default is NULL, in which case the groupname is the directory name supplied by indir,i.e., the result of basename(indir).}
}
\description{
copy STACKS SNPs VCF files using informative names

Copies each of the 'populations.snps.vcf' files generated by the 'populations' program of the STACKS pipeline, and uses a more informative filename. This function will only work if the 'populations_setup' function was used for running populations.
}
