% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_structure.R
\name{create_popfile}
\alias{create_popfile}
\title{Generate a popfile from STRUCTURE output

Takes as input the output files from strucutre runs and generates a pdf with plots of admixture}
\usage{
create_popfile(xdir = NULL, K = 2)
}
\arguments{
\item{xdir}{Path to the directory containing all STRUCTURE output files plus the '.*_sampleIDs.txt' file generated by run_structure. Default NULL. If supplied, this overides save.as, method, labels, and x arguments.}

\item{K}{Number indicating the value of K to use.}
}
\value{
NULL; generates pdf with a barplot of admixture for each K
}
\description{
Generate a popfile from STRUCTURE output

Takes as input the output files from strucutre runs and generates a pdf with plots of admixture
}
\examples{
outdir.temp    <- "PATH/TO/STRUCTURE/OUTPUT"
names.df.path  <- list.files(outdir.temp, full.names=T, pattern="_sampleIDs.txt$")
samplenames.df <- read.table(names.df.path, header=T)
samplenames    <- samplenames.df$IndvNames
# Character vector with paths to output structure files
qfiles         <- list.files(outdir.temp, full.names=T, pattern="log_f$")
admixturePlots(x=qfiles,labels=samplenames,userun=c(1:5))

admixturePlots(xdir="PATH/TO/STRUCTURE/OUTPUT",userun=c(1:5))

}
