% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{cstacks_cleanup}
\alias{cstacks_cleanup}
\title{cleanup cstacks

Deletes cstacks bash scripts in a directory. Run this after all cstacks jobs are completed.}
\usage{
cstacks_cleanup(dir)
}
\arguments{
\item{dir}{Character string with path to directory containing cstacks bash scripts}
}
\description{
cleanup cstacks

Deletes cstacks bash scripts in a directory. Run this after all cstacks jobs are completed.
}
