% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{cstacks_setup}
\alias{cstacks_setup}
\title{batch setup cstacks

For each combination of N and set of ustacks, this function creates a bash job file to run cstacks.
The output of this function is a character string vector with paths to bash job files, which can be piped to the function 'rsbatch' to submit the jobs.}
\usage{
cstacks_setup(save.in, n = 1, p = 16, cstacks.path = NULL)
}
\arguments{
\item{save.in}{Character string with directory where output directories and files should be saved. This should be the same as the value used for the 'save.in' argument of ustacks_setup.}

\item{n}{Number or numerical vector with number of mismatches allowed between sample loci when build the catalog. Default 1.}

\item{p}{Number of threads to use for job parallization. Default 16.}

\item{cstacks.path}{Character string with path to the cstacks executable. Default NULL, in which case the path is derived from the default setting, which can be set with config_miscwrappers('<path/to/cstacks>').}
}
\value{
Character string vector with the paths to bash job files.
}
\description{
batch setup cstacks

For each combination of N and set of ustacks, this function creates a bash job file to run cstacks.
The output of this function is a character string vector with paths to bash job files, which can be piped to the function 'rsbatch' to submit the jobs.
}
