% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPC_adegenet.R
\name{dapc.biplot.arrange}
\alias{dapc.biplot.arrange}
\title{Arrange list of ggplot biplots

Function to arrange ggplot biplots for each pair of DFs or PCs of each K
Max number of plots = 25.}
\usage{
\method{dapc}{biplot.arrange}(
  x,
  variable = "DF",
  layout.mat = NULL,
  row.labels.left = NULL,
  col.labels.top = NULL,
  row.labels.right = NULL,
  col.labels.bottom = NULL,
  pad = 0.1,
  K = NULL
)
}
\arguments{
\item{x}{A list of ggplots produced by ggscatter.dapc function}

\item{variable}{Either "DF" or "PC", or any other character string.}

\item{layout.mat}{Custum layout matrix to use. Default is NULL, in which case the layout will be generated automatically as a fucnton of the number of plots.}

\item{row.labels.left}{Text labels to use to the left of the first column of plots. Default NULL (no labels).}

\item{col.labels.top}{Text labels to use above the first row of plots. Default NULL (no labels).}

\item{row.labels.right}{Text labels to to the right of the last column of plots. Default NULL (no labels).}

\item{col.labels.bottom}{Text labels to use below the bottom row of plots. Default NULL (no labels).}

\item{pad}{Amount of space between plots, in units of line widths (Default 0.1).}

\item{K}{Which set of biplots to use. If NULL (the default), the function will attemp to draw all biplots for all K (max 25 plots).}
}
\value{
A gtable object
}
\description{
Arrange list of ggplot biplots

Function to arrange ggplot biplots for each pair of DFs or PCs of each K
Max number of plots = 25.
}
