% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{decompose_distances}
\alias{decompose_distances}
\title{geo distances

Used for making EEMS plots}
\usage{
decompose_distances(diffs, sizes = NULL)
}
\arguments{
\item{diffs}{Differences}

\item{sizes}{Sizes}
}
\value{
List with parameters
}
\description{
geo distances

Used for making EEMS plots
}
