% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{default_eems_colors}
\alias{default_eems_colors}
\title{default eems colors

Use the default DarkOrange to Blue color scheme, which combines two color schemes from the \code{dichromat} package. These are based
on a collection of color schemes for scientific graphics:
See http://geog.uoregon.edu/datagraphics/color_scales.htm
To reproduce the default eems colors, let oranges be dichromat::colorschemes$BluetoDarkOrange.12\link{12:7}
and blues be dichromat::colorschemes$BrowntoBlue.12\link{7:12}
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').}
\usage{
default_eems_colors()
}
\value{
Character string vector with hex codes for default eems colors
}
\description{
default eems colors

Use the default DarkOrange to Blue color scheme, which combines two color schemes from the \code{dichromat} package. These are based
on a collection of color schemes for scientific graphics:
See http://geog.uoregon.edu/datagraphics/color_scales.htm
To reproduce the default eems colors, let oranges be dichromat::colorschemes$BluetoDarkOrange.12\link{12:7}
and blues be dichromat::colorschemes$BrowntoBlue.12\link{7:12}
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').
}
