% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.outer.R
\name{dfTransform}
\alias{dfTransform}
\title{Transform CRS coordinates data frame

Transform coordinates held in a dataframe given the starting and final coordinate reference systems}
\usage{
dfTransform(df, x, y, CRS1 = "EPSG:4326", CRS2)
}
\arguments{
\item{x}{Name of the column (variable) with longitude}

\item{y}{Name of the column (variable) with latitude}

\item{CRS1}{CRS of the input coordinates}

\item{CRS2}{CRS to use for the output coordinates}

\item{data}{frame with, minimally, longitude and latitude columns}
}
\value{
data frame with reprojected coordinates. Other columns of the input data frame are preserved.
}
\description{
Transform CRS coordinates data frame

Transform coordinates held in a dataframe given the starting and final coordinate reference systems
}
