% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPC_adegenet.R
\name{diffcol}
\alias{diffcol}
\title{Cartesian color distance

Cartesian distance between two colors supplied in hexidecimal}
\usage{
diffcol(col1, col2)
}
\arguments{
\item{col1}{Hex color code for color #1}

\item{col2}{Hex color code for color #2}
}
\value{
Number with cartesional color distance
}
\description{
Cartesian color distance

Cartesian distance between two colors supplied in hexidecimal
}
