% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{dlogunif}
\alias{dlogunif}
\title{dlunif function

loguniform probability density function}
\usage{
dlogunif(x, min, max, base = exp(1))
}
\arguments{
\item{x}{sample quantile}

\item{min}{minimum of sampling distribution}

\item{max}{maximum of sampling distribution}

\item{base}{base of exponent. Default exp(1)}
}
\value{
probability density at x
}
\description{
dlunif function

loguniform probability density function
}
\examples{
source("plot_model.R")
dev.new(width=10,height=6)
Model5.plot  <- plot_model(tpl.path="example_5.tpl",  est.path="example_5.est")
Model6.plot  <- plot_model(tpl.path="example_6.tpl",  est.path="example_6.est")
Model7.plot  <- plot_model(tpl.path="example_7.tpl",  est.path="example_7.est")
Model8.plot  <- plot_model(tpl.path="example_8.tpl",  est.path="example_8.est")
Model9.plot  <- plot_model(tpl.path="example_9.tpl",  est.path="example_9.est")
Model10.plot <- plot_model(tpl.path="example_10.tpl", est.path="example_10.est")
Model11.plot <- plot_model(tpl.path="example_11.tpl", est.path="example_11.est")
Model12.plot <- plot_model(tpl.path="example_12.tpl", est.path="example_12.est")
Model13.plot <- plot_model(tpl.path="example_13.tpl", est.path="example_13.est")
Model14.plot <- plot_model(tpl.path="example_14.tpl", est.path="example_14.est")
Model15.plot <- plot_model(tpl.path="example_15.tpl", est.path="example_15.est")
Model16.plot <- plot_model(tpl.path="example_16.tpl", est.path="example_16.est")
Model17.plot <- plot_model(tpl.path="example_17.tpl", est.path="example_17.est")
Model18.plot <- plot_model(tpl.path="example_18.tpl", est.path="example_18.est")

list.of.models <- list(Model5.plot,Model6.plot,Model7.plot,Model8.plot,Model9.plot,Model10.plot,Model11.plot,Model12.plot,Model13.plot,Model14.plot,Model15.plot,Model16.plot,Model17.plot,Model18.plot)
pdf("cartoon_models.pdf",width=10,height=6)
list.of.models
dev.off()
}
