% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{dunif.quotient.a0}
\alias{dunif.quotient.a0}
\title{Probability density function for the quotient of two uniform distributions that both have minimum value at zero.

Probability density for z in Z ~ X1/X2, where X1 ~ U(0,b1) and X2 ~ U(0,b2), b1 > 0, b2 > 0.}
\usage{
dunif.quotient.a0(z, b)
}
\arguments{
\item{z}{quantile}

\item{b}{Numerical vector with b1 and b2 for X1~Unif(0,b1) and X2~Unif(0,b2).}
}
\value{
Probability density of z in Z
}
\description{
Probability density function for the quotient of two uniform distributions that both have minimum value at zero.

Probability density for z in Z ~ X1/X2, where X1 ~ U(0,b1) and X2 ~ U(0,b2), b1 > 0, b2 > 0.
}
