% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_delimitR.R
\name{easySFS_fsc2}
\alias{easySFS_fsc2}
\title{Run easySFS even easier to generate input MSFS and population files for fastsimcoal2}
\usage{
easySFS_fsc2(easySFS.path, vcf, traits, output.dir)
}
\arguments{
\item{easySFS.path}{Character string with path to the easySFS executable}

\item{vcf}{Character string with path to the VCF file}

\item{traits}{Character string with path to the traits file}

\item{output.dir}{Character string with path to directory where the two output files should be saved.}
}
\description{
Given VCF and traits files usually needed by easySFS, generate the fsc2 MSFS and the pops file needed for fastsimcoal2
Chooses the projection that yields the greatest number of segregating sites, without trying to balance for the number of individuals.
}
\details{
IMPORTANT: You must include the next two lines in the sh file for any job that uses this function. For now this function will only work for me.
module load anaconda/4.7
conda activate easySFS

requires gtools package
}
