% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runeems_snps_setup.R
\name{eems_cleanup}
\alias{eems_cleanup}
\title{cleanup eems

Deletes eems_setup bash and R scripts in a directory. Run this after all eems jobs are completed. Does not delete the "jobX.sh" bash files that run eems.}
\usage{
eems_cleanup(dir)
}
\arguments{
\item{dir}{Character string with path to directory containing eems .sh amd .R scripts}
}
\description{
cleanup eems

Deletes eems_setup bash and R scripts in a directory. Run this after all eems jobs are completed. Does not delete the "jobX.sh" bash files that run eems.
}
\examples{
library(misc.wrappers)

# Path to VCF with SNPs
vcf.path    <- file.path(system.file("extdata", package = "misc.wrappers"), "simK4.vcf.gz")
# Path to file with longitude and latitude of sampling locality of each individual
coords.path <- file.path(system.file("extdata", package = "misc.wrappers"), "simK4_coords.txt")
# Where to save output
save.path   <- file.path(system.file("examples", package = "misc.wrappers"),"simK4")
# Setup environment and input files for runeems_snps
eems.setup  <- runeems_snps_setup(x=vcf.path, coords=coords.path, save.in=save.path, numMCMCIter = 100000, numBurnIter = 10000, numThinIter = 999)
}
