% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg2raster.R
\name{eemsgg2raster}
\alias{eemsgg2raster}
\title{eemsgg2raster function

Convert one of the four maps produced by make_eems_plots from class ggplot to class RasterLayer, and optionally write the raster as a geoTIFF file.}
\usage{
eemsgg2raster(gg.obj, file.out = NULL)
}
\arguments{
\item{file.out}{Path where raster should be saved. Use extension '.tif' to save as geoTIFF.}

\item{x}{One of the plots produced by make_eems_plots: plots$mrates01, plots$mrates02, plots$qrates01, plots$qrates02}
}
\value{
Onject of class RasterLayer
}
\description{
eemsgg2raster function

Convert one of the four maps produced by make_eems_plots from class ggplot to class RasterLayer, and optionally write the raster as a geoTIFF file.
}
