% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmaphelper.R
\name{equivalencyMats}
\alias{equivalencyMats}
\title{ennumerate equivalence relations

A list of n x n matrices, corresponding to all possible equivalence relation for a set with n elements.
WARNING: long compute time for values of n > 7; otherwise fast.}
\usage{
equivalencyMats(n)
}
\arguments{
\item{n}{Number of elements in the unpartitioned set}
}
\value{
A list of transition matrices, for all possible partitions of pairwise transitions into parameters describing transitions.
}
\description{
ennumerate equivalence relations

A list of n x n matrices, corresponding to all possible equivalence relation for a set with n elements.
WARNING: long compute time for values of n > 7; otherwise fast.
}
