% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.outer.R
\name{evalfun}
\alias{evalfun}
\title{Evaluate function name with lists of arguments}
\usage{
evalfun(fun.name, supplied.list = NULL, passed = list(...), ...)
}
\arguments{
\item{fun.name}{Character string with name of the function to be evaluated}

\item{supplied.list}{Arguments and values of fun.name defined explicitely in a list}

\item{passed}{A list with names equal to argument names in the set of names of some arguments with values to use for arguments. This list is typically formed by capturing the "additional arguments to be passed with '...'" in a parent function list(...).}

\item{...}{Arguments and values of fun.name defined explicitely and not together in a list}
}
\value{
Value returned by evaluating the function with 'fun.name' and arguments in lists 'supplied' and 'passed'
}
\description{
This function determines which entries in the lists 'supplied' and 'passed' are arguments of the function 'fun.name', and then evaluates the function with those arguments.
Entries with names that are not arguments of 'fun.name' are ignored, making it possible to include arguments for multiple functions called in a parent function with argument '...'. See examples.
}
