% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_structure.R
\name{EvannoPlots}
\alias{EvannoPlots}
\title{Generate Evanno Method Plots

This function uses functions from pophelper and ggplot2 to generate a pdf with the four Evanno Method Plots.}
\usage{
EvannoPlots(input.dir = getwd(), save.as = "EvannoPlots.pdf")
}
\arguments{
\item{input.dir}{directory with output files of STRUCTURE analysis. Default is the current directory.}

\item{save.as}{Where to save the output PDF. Default is "evannoPlots.pdf" in the current directory.}
}
\value{
g table with gpplots arranged in a 2x2 grid
}
\description{
Generate Evanno Method Plots

This function uses functions from pophelper and ggplot2 to generate a pdf with the four Evanno Method Plots.
}
\examples{
library(misc.wrappers)
}
