% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.base.polygons.R
\name{extract.base.polygons}
\alias{extract.base.polygons}
\title{extract.base.polygons function

Extract all lowest-level polygons of a SpatialPolygonsDataFrame}
\usage{
extract.base.polygons(spdf)
}
\arguments{
\item{spdf}{SpatialPolygonsDataFrame (see sp package)}
}
\value{
SpatialPolygons object with each lowest-level polygon of spdf held as a feature.
}
\description{
extract.base.polygons function

Extract all lowest-level polygons of a SpatialPolygonsDataFrame
}
