% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_delimitR.R
\name{fastsimcoalsims_Par}
\alias{fastsimcoalsims_Par}
\title{fastsimcoalsims_Par

Function to run fastsimcoalsims in parallel. This is modified from Paul Hime's function of the same name. Implemented in the function rundelimitR.
Simulates the site frequency spectrum for a set of input models each defined by their .est and .tpl files.}
\usage{
fastsimcoalsims_Par(
  prefix,
  pathtofsc,
  nreps = 10000,
  wd = getwd(),
  ncores = 2,
  run.parallel = T
)
}
\arguments{
\item{prefix}{Prefix used for .tpl and .est model files}

\item{pathtofsc}{Path to the fastsimcoal2 executable}

\item{nreps}{Number of replicate runs to do for each model. Default is 10000.}

\item{wd}{Working directory. Default is to use the current working directory.}

\item{ncores}{Number of cores to use}

\item{run.parallel}{If simulations on different models should also be run in parallel. Default is TRUE. In some cases running in parallel causes weird issues and in such cases try changing this to FALSE.}
}
\value{
NULL. Results are written to working directory.
}
\description{
fastsimcoalsims_Par

Function to run fastsimcoalsims in parallel. This is modified from Paul Hime's function of the same name. Implemented in the function rundelimitR.
Simulates the site frequency spectrum for a set of input models each defined by their .est and .tpl files.
}
\examples{

run_delimitR(vcf="PATH/TO/VCF",traitsfile="PATH/TO/TRAITSFILE")
}
