% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{filled_contour_rates}
\alias{filled_contour_rates}
\title{filled contour rates

Generates a ggplot object with filled contours
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').}
\usage{
filled_contour_rates(z, dimns)
}
\arguments{
\item{z}{}

\item{dimns}{output of read_dimns function}
}
\value{
ggplot object
}
\description{
filled contour rates

Generates a ggplot object with filled contours
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').
}
