% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{filled_eems_contour}
\alias{filled_eems_contour}
\title{filled eems contour

generates a ggplot object with filled contours
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').}
\usage{
filled_eems_contour(dimns, zmean, plot_params, is_mrates)
}
\arguments{
\item{dimns}{output of read_dimns function}

\item{zmean}{vector of zeros with length determined by the named entry 'nmrks' of the output of read_dimns function}

\item{plot_params}{List of plotting parameters}

\item{is_mrates}{Logical indicating if the plots are mrates, else qrates}
}
\value{
ggplot object
}
\description{
filled eems contour

generates a ggplot object with filled contours
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').
}
