% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{filled_prob_contour}
\alias{filled_prob_contour}
\title{filled probability contour

Generates asthetics for ggplots.
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').}
\usage{
filled_prob_contour(dimns, probs, plot_params, is_mrates)
}
\arguments{
\item{dimns}{output of read_dimns function}

\item{probs}{Probabilities}

\item{plot_params}{List of plot parameters}

\item{is_mrates}{Logical indicating if the plots are mrates, else qrates}
}
\value{
ggplot aesthetics
}
\description{
filled probability contour

Generates asthetics for ggplots.
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').
}
