% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmaphelper.R
\name{fitmodels}
\alias{fitmodels}
\title{Fit data and a list of Mk models with simmap

Makes a list of all possible model schemes for transitions between a set of states.}
\usage{
fitmodels(tree, priors.tips, model.transitions, priors.root = NULL)
}
\arguments{
\item{tree}{newick character string with phylogenetic tree}

\item{priors.tips}{matrix with character state priors for the tips of the tree}

\item{model.transitions}{a square matrix or list of square matrices for the transition rate parameter schemes; a list with all possible parameter schemes can be generated with the function transitionModels}

\item{priors.root}{a numeric vector with the priors for the character states at the root node. Default is a uniform distribution.}
}
\value{
A list of transition matrices, for all possible partitions of pairwise transitions into parameters describing transitions.
}
\description{
Fit data and a list of Mk models with simmap

Makes a list of all possible model schemes for transitions between a set of states.
}
