% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracD.R
\name{fracD}
\alias{fracD}
\title{fracD function

Calculates the fractal dimension of perimeter-area ratio, a metric of polygon complexity.
More info: http://www.umass.edu/landeco/teaching/landscape_ecology/schedule/chapter9_metrics.pdf}
\usage{
fracD(x)
}
\arguments{
\item{x}{SpatialPolygons object (see sp package)}
}
\value{
Number in [1,2), with larger values returned for increasing complex polygons; a value of 1 is returned for simple euclidean shapes likes squares and circles.
}
\description{
fracD function

Calculates the fractal dimension of perimeter-area ratio, a metric of polygon complexity.
More info: http://www.umass.edu/landeco/teaching/landscape_ecology/schedule/chapter9_metrics.pdf
}
