% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genind2diffs.R
\name{genind2diffs}
\alias{genind2diffs}
\title{genind2diffs function

Generates a diffs matrix from a genind object}
\usage{
genind2diffs(
  genind.obj,
  ploidy = 2,
  include.indv.names = F,
  output.file = NULL
)
}
\arguments{
\item{genind.obj}{Input genind object (see adegenet package for description of genind class). Non-biallelic sites are dropped if they exist.}

\item{ploidy}{Integer indicating the ploidy of individuals}

\item{include.indv.names}{If the output matrix should include the names of the individuals as rownames and column names. Default FALSE}

\item{output.file}{Optional character string to save the result to a file.}
}
\value{
A list of length three   with (1) a numeric matrix corresponding to the Diff matrix used for EEMS, (2) the number of individuals, (3) number of biallelic sites in genind.obj
}
\description{
genind2diffs function

Generates a diffs matrix from a genind object
}
