% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{geo_distm}
\alias{geo_distm}
\title{geo distances

Used for making EEMS plots
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').}
\usage{
geo_distm(coord, longlat, plot_params)
}
\arguments{
\item{coord}{input coordinates}

\item{longlat}{Whether or not to use long lat coordinates. Default TRUE}

\item{plot_params}{List of plotting parameters}
}
\value{
Distance between two points
}
\description{
geo distances

Used for making EEMS plots
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').
}
