% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{get_dist_metric}
\alias{get_dist_metric}
\title{get distance metric

Get the distance metric from \code{eemsrun.txt}; use \code{euclidean} by default.
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').}
\usage{
get_dist_metric(mcmcpath)
}
\arguments{
\item{mcmcpath}{Path the directory containing EEMs output files for a particular chain}
}
\value{
Character string with distance metric to use
}
\description{
get distance metric

Get the distance metric from \code{eemsrun.txt}; use \code{euclidean} by default.
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').
}
