% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.outer.R
\name{gg.world110_proj}
\alias{gg.world110_proj}
\title{ggplot projected Natural Earth

This function creates a ggplot object of Earth from the 110m resolution Natural Earth dataset, and the function is useful for exploring different map projections.}
\usage{
gg.world110_proj(CRS = "EPSG:4326", df = NULL)
}
\arguments{
\item{CRS}{Character string with name of EPSG or ESRI coordinate reference system ID. Default is 'EPSG:4326'.}

\item{df}{Optional data frame with longitude and latitude coordinates (datum must be WGS84) in first two columns, respectively, for points that should be plotted on the map.}
}
\value{
ggplot object of Natural Earth data plotted using the projection specified by 'CRS' argument.
}
\description{
ggplot projected Natural Earth

This function creates a ggplot object of Earth from the 110m resolution Natural Earth dataset, and the function is useful for exploring different map projections.
}
