% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPC_adegenet.R
\name{ggscatter.dapc}
\alias{ggscatter.dapc}
\title{Make ggplot with DAPC scatterplot

This function produces plots very similar to those produced by the adegenet function 'scatter.dapc', except that the object returned is a ggplot. This is useful when making lists of plots.
SOME FEATURES STILL IN PROGRESS:
\itemize{
\item ability to add screeplot of discriminant functions or PCAs.
}}
\usage{
ggscatter.dapc(
  x,
  xax = 1,
  yax = 2,
  grp = x$grp,
  cpoint = 2,
  col = adegenet::seasun(length(levels(grp))),
  txt.leg = levels(grp),
  label = levels(grp),
  pch = 20,
  solid = 0.7,
  scree.da = TRUE,
  scree.pca = FALSE,
  posi.da = "bottomright",
  posi.pca = "bottomleft",
  bg.inset = "white",
  ratio.da = 0.25,
  ratio.pca = 0.25,
  inset.da = 0.02,
  inset.pca = 0.02,
  inset.solid = 0.5,
  onedim.filled = TRUE,
  mstree = FALSE,
  lwd = 1,
  lty = 1,
  segcol = "black",
  legend = FALSE,
  posi.leg = "topright",
  cleg = 1,
  cstar = 1,
  cellipse = 1.5,
  axesell = FALSE,
  clabel = 1,
  xlim = NULL,
  ylim = NULL,
  grid = FALSE,
  addaxes = TRUE,
  origin = c(0, 0),
  include.origin = TRUE,
  sub = "",
  csub = 1,
  possub = "bottomleft",
  cgrid = 1,
  pixmap = NULL,
  contour = NULL,
  area = NULL,
  label.inds = NULL,
  new.pred = NULL
)
}
\arguments{
\item{x}{Object of class DAPC (adegenet package).}

\item{xax}{Which descriminant function to plot on the x axis.}

\item{yax}{Which descriminant function to plot on the y axis. Default 2. Ignored if only one discriminant function exists.}

\item{grp}{Object of class 'factor' with length equal to the number of individuals, which indicates individual assignments to clusters. By default posterior assignments are extracted from 'x'.}

\item{cpoint}{Number with cex size to use for points.}

\item{col}{Vector with color to use for each cluster.}

\item{txt.leg}{NOT YET IMPLEMENTED. Labels to use for clusters in the legend.}

\item{label}{Name to use for clusters. Default NULL, in which case group will be used.}

\item{pch}{Either one number indicating pch code of symbol to use for all points (see 'pch' in graphical parameters), or, a numeric vector with pch codes to each for clusters . Default 20.}

\item{solid}{Number between 0 and 1 indicating the level of transparency to use for colors; 0 = fully transparent; 1 = fully opaque; default = 0.7.}

\item{scree.da}{NOT YET IMPLEMENTED. Logical indicated whether or not a screeplot of the discriminant functions should be included. Default TRUE.}

\item{scree.pca}{NOT YET IMPLEMENTED. Logical indicated whether or not a screeplot of the PCs should be included. Default FALSE.}

\item{posi.da}{NOT YET IMPLEMENTED. The position of the screeplot of discriminant functions. Can match any combination of "top/bottom" and "left/right", or a set of x/y coordinates stored as a list (locator can be used). Default = "bottomright".}

\item{posi.pca}{NOT YET IMPLEMENTED. The position of the screeplot of discriminant functions. Can match any combination of "top/bottom" and "left/right", or a set of x/y coordinates stored as a list (locator can be used). Default = "bottomleft".}

\item{bg.inset}{NOT YET IMPLEMENTED. Default "white".}

\item{ratio.da}{NOT YET IMPLEMENTED. Default 0.25}

\item{ratio.pca}{NOT YET IMPLEMENTED. Default 0.25}

\item{inset.da}{NOT YET IMPLEMENTED. Default 0.02}

\item{inset.pca}{NOT YET IMPLEMENTED. Default 0.02}

\item{inset.solid}{NOT YET IMPLEMENTED. Default 0.5}

\item{onedim.filled}{Logical indicating, when only one discriminant function is to be plotted, whether or not density plots should be filled or unfilled with the colors indicated by 'col'. Default TRUE.}

\item{mstree}{A logical indicating whether a minimum spanning tree linking the groups and based on the squared distances between the groups inside the entire space should added to the plot (TRUE), or not (FALSE). Default FALSE.}

\item{lwd}{Line weight to use for edges of the minimum spanning tree linking the groups. Default 0.5.}

\item{lty}{Line type to use for edges of the minimum spanning tree linking the groups. Default 1 (solid).}

\item{segcol}{Color to use for edges of the minimum spanning tree linking the groups Default "black".}

\item{legend}{A logical indicating whether a legend for group colours should added to the plot. Default FALSE.}

\item{posi.leg}{NOT YET IMPLEMENTED. The position of the legend holding group colors. Can match any combination of "top/bottom" and "left/right", or a set of x/y coordinates stored as a list (locator can be used). Default = "topright".}

\item{cleg}{NOT YET IMPLEMENTED. Size factor for the legend. Default 1.}

\item{cstar}{A number greater than 0 defining the length of the star size (i.e., the lines radiating from the center of clusters towards individuals). Default 1. Setting to zero removes the star lines; setting =1 connects points to cluster mean; setting > 1 extends lines through their points.}

\item{cellipse}{A positive coefficient for the inertia ellipse size. Default 1.5. Setting to zero removes ellipses.}

\item{axesell}{NOT YET IMPLEMENTED. A logical value indicating whether the ellipse axes should be drawn. Default FALSE.}

\item{clabel}{A number indicating the size of labels. Default 1.}

\item{xlim}{NOT YET IMPLEMENTED. Default NULL.}

\item{ylim}{NOT YET IMPLEMENTED. Default NULL.}

\item{grid}{NOT YET IMPLEMENTED. Whether or not to include a grid in the background. Default FALSE.}

\item{addaxes}{Logical indicating if reference axes should be drawn at x=a and y=b, with a and b supplied by the 'origin' argument. Default TRUE.}

\item{origin}{NOT YET IMPLEMENTED. Location of the origin.}

\item{include.origin}{NOT YET IMPLEMENTED. A logical value indicating whether the point "origin" should be belonged to the graph space. Default TRUE.}

\item{sub}{NOT YET IMPLEMENTED. A string of characters to be inserted as legend. Default "".}

\item{csub}{NOT YET IMPLEMENTED. Number specifying text size for 'sub'.}

\item{possub}{NOT YET IMPLEMENTED. The position of the subtitle ("topleft", "topright", "bottomleft", "bottomright"). Default "bottomleft".}

\item{cgrid}{NOT YET IMPLEMENTED. A number used with par("cex")* cgrid to specify the mesh of the grid.}

\item{pixmap}{NOT YET IMPLEMENTED. An object 'pixmap' displayed in the map background. Default NULL.}

\item{contour}{NOT YET IMPLEMENTED. A data frame with 4 columns to plot the contour of the map : each row gives a segment (x1,y1,x2,y2). Default NULL.}

\item{area}{NOT YET IMPLEMENTED. A data frame of class 'area' to plot a set of surface units in contour. Default NULL.}

\item{label.inds}{NOT YET IMPLEMENTED. Default NULL. Named list of arguments passed to the orditorp function. This will label individual points witout overlapping. Arguments x and display are hardcoded and should not be specified by user.}

\item{new.pred}{NOT YET IMPLEMENTED. An optional list, as returned by the predict method for dapc objects; if provided, the individuals with unknown groups are added at the bottom of the plot. To visualize these individuals only, specify only.grp="unknown".}

\item{bg}{Color to use for the background of the plot. Default "white".}

\item{only.grp}{NOT YET IMPLEMENTED. Character vector indicating which groups should be displayed. Values should match values of x$grp. If NULL, all results are displayed.}
}
\value{
A ggplot object
}
\description{
Make ggplot with DAPC scatterplot

This function produces plots very similar to those produced by the adegenet function 'scatter.dapc', except that the object returned is a ggplot. This is useful when making lists of plots.
SOME FEATURES STILL IN PROGRESS:
\itemize{
\item ability to add screeplot of discriminant functions or PCAs.
}
}
