% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPC_adegenet.R
\name{goodcolors2}
\alias{goodcolors2}
\title{Return my favorite set of colors to use for clusters in admixture plots, for a particular value of K.

For each K (2-15), a particular result from the 'goodcolors' function that I like.}
\usage{
goodcolors2(n, plot.palette = FALSE)
}
\arguments{
\item{n}{Number of colors to include in the output set. Equal to K if using this function for population structure analyses that produce admixture plots or population assignment probabilities.}

\item{plot.palette}{Whether or not to plot the color palette in the plotting window. Default FALSE.}
}
\description{
Return my favorite set of colors to use for clusters in admixture plots, for a particular value of K.

For each K (2-15), a particular result from the 'goodcolors' function that I like.
}
