% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{gstacks_cleanup}
\alias{gstacks_cleanup}
\title{cleanup gstacks

Deletes gstacks bash scripts in a directory. Run this after all gstacks jobs are completed.}
\usage{
gstacks_cleanup(dir)
}
\arguments{
\item{dir}{Character string with path to directory containing gstacks bash scripts}
}
\description{
cleanup gstacks

Deletes gstacks bash scripts in a directory. Run this after all gstacks jobs are completed.
}
