% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{gstacks_setup}
\alias{gstacks_setup}
\title{batch setup gstacks

For each set samples (processed with ustacks, cstacks, sstacks, and tsv2bam) and a population map, this function creates a bash job file to run gstacks.
The output of this function is a character string vector with paths to bash job files, which can be piped to the function 'rsbatch' to submit the jobs.}
\usage{
gstacks_setup(save.in, popmap = NULL, t = 16, gstacks.path = NULL)
}
\arguments{
\item{save.in}{Character string with directory where output directories and files should be saved. This should be the same as the value used for the 'save.in' argument of ustacks_setup.}

\item{popmap}{Character string to population map. Default NULL, in which case all individuals in 'save.in' are assigned to the same 'population'.}

\item{t}{Number of threads to use for job parallization. Default 16.}

\item{gstacks.path}{Character string with path to the gstacks executable. Default NULL, in which case the path is derived from the default setting, which can be set with config_miscwrappers('<path/to/gstacks>').}
}
\value{
Character string vector with the paths to bash job files.
}
\description{
batch setup gstacks

For each set samples (processed with ustacks, cstacks, sstacks, and tsv2bam) and a population map, this function creates a bash job file to run gstacks.
The output of this function is a character string vector with paths to bash job files, which can be piped to the function 'rsbatch' to submit the jobs.
}
