% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{is_color}
\alias{is_color}
\title{Is color

Check that a string, or a vector of strings, represents a hex color
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').}
\usage{
is_color(x)
}
\arguments{
\item{x}{Character string to check}
}
\value{
Logical
}
\description{
Is color

Check that a string, or a vector of strings, represents a hex color
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').
}
