% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minCladesList.R
\name{minCladesList}
\alias{minCladesList}
\title{minimum length set of clades with tip labels

Returns a smallest set of non-intersecting, non-sister clades of an input tree, and which together contain all of the tip labels supplied in the vector tiplabs.}
\usage{
minCladesList(tree, tiplabs, quiet = T, warn = T)
}
\arguments{
\item{tree}{A tree object of class phylo or treedata.}

\item{tiplabs}{A character vector with a subset of the tip labels of the input tree}

\item{quiet}{Logical indicating whether to suppress some messages printed to screen. Default TRUE. This does not affect warnings.}

\item{warn}{Logical indicating if a warning should be generated when one or more tiplabs do not occur in the input tree. Default TRUE.}
}
\value{
A list of character vectors, each containing a subset of the names supplied to tiplabs, and each characterizing a clade that is non-intersecting and non-sister to any other clade in the list.
}
\description{
minimum length set of clades with tip labels

Returns a smallest set of non-intersecting, non-sister clades of an input tree, and which together contain all of the tip labels supplied in the vector tiplabs.
}
