% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPC_adegenet.R
\name{newpoint}
\alias{newpoint}
\title{Coordinates of point on ray

Used in the function ggscatter.dapc to control segments radiating from cluster centers towards individual's pca coordinates.
Returns x,y coordinates of a point (p2) on a ray with origin at p0(x0,y0) and another point at p1(x1,y1), such that the distance from p0 to p2 = c*(distance from p0 to p1).}
\usage{
newpoint(p0, p1, c)
}
\arguments{
\item{p0}{Numberical vector with coordinates c(x0,y0) defining the origin of a ray}

\item{p1}{Numberical vector with coordinates c(x1,y1) defining a point on the ray other than the origin.}

\item{c}{A number equal to (distance p0 to p2)/(distance p0 to p1)}
}
\value{
Numberical vector with coordinates c(x2,y2) of p2.
}
\description{
Coordinates of point on ray

Used in the function ggscatter.dapc to control segments radiating from cluster centers towards individual's pca coordinates.
Returns x,y coordinates of a point (p2) on a ray with origin at p0(x0,y0) and another point at p1(x1,y1), such that the distance from p0 to p2 = c*(distance from p0 to p1).
}
