% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{pairwise_dist}
\alias{pairwise_dist}
\title{pairwise distances between demes

Internal function used for making EEMS plots (plots 5–7). Calculates pairwise distances between demes.
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').}
\usage{
pairwise_dist(mcmcpath, longlat, plot_params)
}
\arguments{
\item{mcmcpath}{Path the directory containing EEMs output files for a particular chain}

\item{longlat}{Whether or not to use long lat coordinates. Default TRUE}

\item{plot_params}{List of plotting parameters}
}
\value{
List with parameter values used for the function plot_pairwise_dissimilarities_
}
\description{
pairwise distances between demes

Internal function used for making EEMS plots (plots 5–7). Calculates pairwise distances between demes.
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').
}
