% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot.funUnif}
\alias{plot.funUnif}
\title{plot PDF for sum or product of two uniform distributions

Plots a line chart for the probability density distribution (PDF) or cumulative distribution (CDF) for the sum or product of uniform distributions on positive Reals.
Can handle an arbitrary number of input distributions, which must be independent, and can have different ranges.}
\usage{
\method{plot}{funUnif}(
  a,
  b,
  FUN = "sum",
  cumulative = F,
  buffer = 0.1,
  nZ = 500,
  show.xy.pd = F,
  line.col = c("red"),
  line.lty = c(1),
  add = F,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{a}{A numerical vector holding the minimum values of input uniform distributions.}

\item{b}{A numerical vector holding the maximum values of input uniform distributions.}

\item{FUN}{The function to use to determine PDF or CDF; one of 'simple', 'sum', or 'product'. If a and b are length 1 vectors, 'simple' is used.}

\item{cumulative}{If TRUE, the CDF will be plotted instead of the PDF}

\item{buffer}{fraction of extra, zero-probability region along x axis to plot}

\item{show.xy.pd}{When FUN = 'sum' or 'product', if 'show.xy.pd' = TRUE the pdf or cdf of component distributions will be plotted along with their sum or product.}

\item{line.col}{color of curve}

\item{line.lty}{type of curve}

\item{add}{Whether or not to the graph should be added to the existing plot. Default is FALSE.}

\item{xlim}{min and max limits for the x-axis plotting area}

\item{ylim}{min and max limits for the y-axis plotting area}
}
\value{
A line graph showing the PDF of the sum or product of the two input uniform distributions
}
\description{
plot PDF for sum or product of two uniform distributions

Plots a line chart for the probability density distribution (PDF) or cumulative distribution (CDF) for the sum or product of uniform distributions on positive Reals.
Can handle an arbitrary number of input distributions, which must be independent, and can have different ranges.
}
