% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{plot_eems}
\alias{plot_eems}
\title{plot_eems function

Invokes the reemsplots2 function make_eems_plots, with additional functionality such as adding island borders and points showing individual coordinates..}
\usage{
plot_eems(
  mcmcdir,
  output.plot.path = NULL,
  plot.coords = T,
  plot.geography = T,
  plot.oceans = T
)
}
\arguments{
\item{mcmcdir}{Character string to directory of a particular mcmc chain}

\item{output.plot.path}{Character string with path where to save output. If NULL, the plot is printed but not saved.}

\item{plot.coords}{Whether or not the sample}
}
\value{
NULL; plots are printed to screen and saved to output.plot.path
}
\description{
plot_eems function

Invokes the reemsplots2 function make_eems_plots, with additional functionality such as adding island borders and points showing individual coordinates..
}
