% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{plot_log_posterior}
\alias{plot_log_posterior}
\title{plot eems log posterior

Used for making EEMS plot 8}
\usage{
plot_log_posterior(mcmcpath)
}
\arguments{
\item{mcmcpath}{Path the directory containing EEMs output files for a particular chain}
}
\value{
Object of class ggplot
}
\description{
plot eems log posterior

Used for making EEMS plot 8
}
