% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{plot_pairwise_dissimilarities_}
\alias{plot_pairwise_dissimilarities_}
\title{plot pairwise dissimilarities between demes

Internal function used for making EEMS plots (plots 5–7).}
\usage{
plot_pairwise_dissimilarities_(dissimilarities, add_abline)
}
\arguments{
\item{dissimilarities}{Output list from the function pairwise_dist}

\item{add_abline}{Whether or not to add the abline.}
}
\value{
List of ggplot objects
}
\description{
plot pairwise dissimilarities between demes

Internal function used for making EEMS plots (plots 5–7).
}
