% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{populations_cleanup}
\alias{populations_cleanup}
\title{cleanup populations

Deletes populations bash scripts in a directory. Run this after all populations jobs are completed.}
\usage{
populations_cleanup(dir)
}
\arguments{
\item{dir}{Character string with path to directory containing populations bash scripts}
}
\description{
cleanup populations

Deletes populations bash scripts in a directory. Run this after all populations jobs are completed.
}
