% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{produnif}
\alias{produnif}
\title{PDF or CDF for the product of n independent uniformly continuous random variables.

Evaluates the probability density function or cumulative distribution of Z=z for Z(z) = Xi*...*Xn, for i = 1...n, Xi ~ U(ai,bi), where 0 < ai < bi
The PDF and CDF is calculated using Equations 3.4 and 3.9, respectively, from Ishihara, Tatsuo. 2002. The Distribution of the Sum and the Product of Independent Uniform Random Variables Distributed at Different Intervals. Transactions - Japan Society for Industrial and Applied Mathematics, 12(3), 197-208.}
\usage{
produnif(z, a, b, cumulative = F)
}
\arguments{
\item{z}{Number or numerical vector of quantiles}

\item{a}{Vector holding the mins of each X}

\item{b}{Vector holding the max of each X}

\item{cumulative}{Whether or not return the cumulative probability instead of the probability. Default is FALSE. This parameter does not yet calculate CDF correctly.}
}
\value{
Number or numerical vector with either the probability density or the cumulative density for each input value z.
}
\description{
PDF or CDF for the product of n independent uniformly continuous random variables.

Evaluates the probability density function or cumulative distribution of Z=z for Z(z) = Xi*...*Xn, for i = 1...n, Xi ~ U(ai,bi), where 0 < ai < bi
The PDF and CDF is calculated using Equations 3.4 and 3.9, respectively, from Ishihara, Tatsuo. 2002. The Distribution of the Sum and the Product of Independent Uniform Random Variables Distributed at Different Intervals. Transactions - Japan Society for Industrial and Applied Mathematics, 12(3), 197-208.
}
