% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{read_dimns}
\alias{read_dimns}
\title{Read dimns

Internal function used for making EEMS plots.
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').}
\usage{
read_dimns(mcmcpath, longlat = TRUE, nmrks = 100)
}
\arguments{
\item{mcmcpath}{Path the directory containing EEMs output files for a particular chain}

\item{longlat}{Whether or not to use long lat coordinates. Default TRUE}

\item{nmrks}{Number of marks. Default 100.}
}
\value{
A list with parameters to use for generating EEMS plots.
}
\description{
Read dimns

Internal function used for making EEMS plots.
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').
}
