% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{read_vector}
\alias{read_vector}
\title{read vector

Internal function used for making EEMS plots. Reads eems output file as a vector.}
\usage{
read_vector(file)
}
\arguments{
\item{file}{Path to file to read}
}
\value{
Vector with file contents
}
\description{
read vector

Internal function used for making EEMS plots. Reads eems output file as a vector.
}
