% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{read_voronoi}
\alias{read_voronoi}
\title{read voronoi

Reads EEMs output and constructs parameters for voronoi interpolation}
\usage{
read_voronoi(mcmcpath, longlat, is_mrates)
}
\arguments{
\item{mcmcpath}{Path the directory containing EEMs output files for a particular chain}

\item{longlat}{Whether or not to use long lat coordinates. Default TRUE}

\item{is_mrates}{Logical indicating if the plots are mrates, else qrates}
}
\value{
List with parameter values used for veronoi interpolation
}
\description{
read voronoi

Reads EEMs output and constructs parameters for voronoi interpolation
}
