% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPC_adegenet.R
\name{rgb.cb}
\alias{rgb.cb}
\title{Transform sRGB colorspace into colorblind colorspaces

Uses the model described here: https://ixora.io/projects/colorblindness/color-blindness-simulation-research/}
\usage{
rgb.cb(srgb)
}
\arguments{
\item{srgb}{Numerical vector with R,G,B color coordinates}
}
\value{
Numeric matrix with transformed RGB color coordinates. First row contains input coordinates. Second through fourth rows contain coordinates for Protanopia, Deuteranopia, and Tritanopia colors.
}
\description{
Transform sRGB colorspace into colorblind colorspaces

Uses the model described here: https://ixora.io/projects/colorblindness/color-blindness-simulation-research/
}
