% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.outer.R
\name{rlandcoords}
\alias{rlandcoords}
\title{Random land-intersecting-points in random-centered area.

Returns a set of points sampled from a region with a randomly generated center and specified longitudinal and latitudinal diameters. Points are sampled until a specified number of land-intersecting points are returned.}
\usage{
rlandcoords(lond, latd, size, return.as = "data.frame")
}
\arguments{
\item{lond}{Longitudinal diameter of area}

\item{latd}{Latitudinal diameter of area}

\item{size}{Number of points to return}

\item{return.as}{Character string with class to use for object returned. Default "data.frame". Can also be "matrix" or "SP" (SpatialPoints).}
}
\value{
An object with class equal to the value of 'return.as', which contains the subset of input points that occur on land, or more specifically, not in an ocean.
}
\description{
Random land-intersecting-points in random-centered area.

Returns a set of points sampled from a region with a randomly generated center and specified longitudinal and latitudinal diameters. Points are sampled until a specified number of land-intersecting points are returned.
}
