% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{rlogunif}
\alias{rlogunif}
\title{Sample from loguniform distribution}
\usage{
rlogunif(n, min, max, base = exp(1))
}
\arguments{
\item{n}{Number of samples to draw}

\item{min}{minimum of sampling distribution}

\item{max}{maximum of sampling distribution}

\item{base}{base of exponential parameter. Default exp(1).}
}
\description{
Sample from loguniform distribution
}
