% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_miscwrappers.R
\name{rsbatch}
\alias{rsbatch}
\title{sbatch setup and submit

Creates slurm sbatch submission string, and optionally submits the job using system}
\usage{
rsbatch(
  sh.path,
  partition = NA,
  nodes = 1,
  ntasksPerNode = 1,
  memGb = 50,
  time = c(0, 6, 0, 0),
  submit = FALSE
)
}
\arguments{
\item{sh.path}{Character string with path to bash script}

\item{partition}{Character string with name of partition to submit the job to. Passed to sbatch argument '--partition'. Default NA.}

\item{nodes}{Number of nodes to split the job across. Default 1.}

\item{ntasksPerNode}{Number of tasks (cores) to use per node. Default 1.}

\item{memGb}{Total amount of memory to allocate for the job. Default 50Gb.}

\item{time}{Numerical vector with length 4, with the number of days, hours, minutes, and seconds, respectively, to allocate for the job.}

\item{submit}{Logical with whether or not to submit the jobs to the cluster.}
}
\value{
Either vector of character strings that can be piped to system, or NULL
}
\description{
sbatch setup and submit

Creates slurm sbatch submission string, and optionally submits the job using system
}
