% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPC_adegenet.R
\name{run_DAPC}
\alias{run_DAPC}
\title{Run DAPC and Plot

Run adagenet DAPC analyses and generate plots of BIC vs number of clusters, baplot of alpha optimum number of principle components at each K, admixture for each K, assignments at each K.
If sample coordinates are supplied, this function interpolates cluster membership probabilities on a map, using functions from tess3r package.}
\usage{
run_DAPC(
  vcf,
  kmax = 50,
  coords = NULL,
  reps = 100,
  probs.out = NULL,
  save.as = NULL
)
}
\arguments{
\item{vcf}{Character string with path to vcf file containing snp data}

\item{kmax}{Number indicating the maximum number of clusters to evaluate. Default is NULL, in which case kmax is set to one less than the number of individuals.}

\item{coords}{Optional character string with path to a table with longitude and latitude of individuals in the vcf file, or a matrix or data frame with longitude and latitude columns. Default is NULL, in which case membership probabilities are not interpolated onto a map.}

\item{reps}{Number indicating the number of replicates of 'find.clusters'. Default 100.}

\item{probs.out}{NULL or a character string with location where to write a table containing the membership probabilities for the best K and alpha-optimized number of PCAs.}

\item{save.as}{Character string with where to save the output PDF with plots of results.}
}
\value{
A list of plots.
}
\description{
Run DAPC and Plot

Run adagenet DAPC analyses and generate plots of BIC vs number of clusters, baplot of alpha optimum number of principle components at each K, admixture for each K, assignments at each K.
If sample coordinates are supplied, this function interpolates cluster membership probabilities on a map, using functions from tess3r package.
}
