% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fastStructure.R
\name{run_fastStructure}
\alias{run_fastStructure}
\title{Run run_fastStructure from SNP data in a VCF file and and plot results.

This fucntion is a wrapper that enables running fastStructure with SNP data in either a VCF file or vcfR object and coordinates in file or a matrix or data frame object.

Notes for running fastStructure:
Requires python 2
Cross validation with the "cv.iter" argument not implemented because it crashes my copy of fastStructure, probably due to a python build problem on my end}
\usage{
run_fastStructure(
  vcf,
  coords = NULL,
  kmax = 40,
  reps = 100,
  save.as = NULL,
  cv.iter = 0,
  python.path = NULL,
  fastStructure.path = NULL
)
}
\arguments{
\item{vcf}{Path to input VCF file with SNP data.}

\item{coords}{Either a character string with path to file containing coordinates (longitude in first column, latitude in second column), or matrix object with longitude and latitude columns.}

\item{kmax}{Numerical vector with set of values to use for K. Default 40.}

\item{reps}{Number of repititions. Default 100.}

\item{save.as}{Where to save the output PDF. Default is NULL. **This argument is ignored in some environments. Instead, use dev.new(file="Where/To/Save/Output.pdf",height=6,width=10,noRStudioGD=TRUE) before calling 'runtess'. Then dev.off().}

\item{cv.iter}{Number of iterations to perform for cross-validation. Default is zero. Currently ignored.}

\item{python.path}{Character string with path to python 2 with fastStructure dependencies Numpy, Scipy, Cython, GNU Scientific Library}

\item{fastStructure.path}{Character string with path to folder containing the fastStructure python executable called 'structure.py'}
}
\value{
List of plots
}
\description{
Run run_fastStructure from SNP data in a VCF file and and plot results.

This fucntion is a wrapper that enables running fastStructure with SNP data in either a VCF file or vcfR object and coordinates in file or a matrix or data frame object.

Notes for running fastStructure:
Requires python 2
Cross validation with the "cv.iter" argument not implemented because it crashes my copy of fastStructure, probably due to a python build problem on my end
}
\examples{
vcf2fastStructure(vcf="Ahaetulla-prasina_AllPops_BestSNP.vcf",out="example.str")
}
