% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SNMF.R
\name{run_SNMF}
\alias{run_SNMF}
\title{Wrapper for LEA snmf

Runs LEA function snmf function from VCF input file and optionally interpolates cluster assignments onto a map (if 'coords' argument is non-NULL).}
\usage{
run_SNMF(
  vcf,
  coords = NULL,
  Krange = 1:40,
  reps = 100,
  entropy = TRUE,
  project = "new",
  iter = 500,
  CPU = 2,
  save.as = NULL
)
}
\arguments{
\item{vcf}{Path to vcf input file}

\item{coords}{Default NULL}

\item{Krange}{Default 1:40}

\item{reps}{Default 100}

\item{entropy}{Default TRUE}

\item{project}{Defualt 'new'}

\item{iter}{Default 500}

\item{CPU}{Defualt 2}

\item{save.as}{Path where output PDF should be saved. Default NULL.}
}
\value{
List of plots
}
\description{
Wrapper for LEA snmf

Runs LEA function snmf function from VCF input file and optionally interpolates cluster assignments onto a map (if 'coords' argument is non-NULL).
}
