% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEEMs_snps.R
\name{runEEMs_snps}
\alias{runEEMs_snps}
\title{runEEMs_snps function

Wrapper function for running EEMS. Optionally generates the '\emph{.outer' file from the '}.coords' file automatically.}
\usage{
runEEMs_snps(
  exe.path = NULL,
  input.data,
  coord,
  outer = NULL,
  ask.use.outer = T,
  output.dirpath,
  n.sites,
  ploidy = 2,
  nDemes = 300,
  numMCMCIter = 1e+07,
  numBurnIter = 1e+06,
  numThinIter = 9999,
  nchains = 3,
  setup.only = T
)
}
\arguments{
\item{exe.path}{Character string with path to runeems_snps executable, or NULL (the default) if you've defined the EEMS path with the function config_miscwrappers. Use paths_misc.wrappers() to check if this path has been set.}

\item{input.data}{An object of class genind or vcfR, or a character string with path to a diffs file.}

\item{coord}{Character string with path to coordinates file, which has two columns with longitude and latitude coordinates (decimal degree format) of individuals in the genind object. Columns should be space separated.}

\item{outer}{Either NULL (the default) or a character string with path to the '\emph{.outer' file, which has two columns with longitude and latitude coordinates (decimal degree format) defining the perimeter of a polygon covering the region to model with EEMS. See EEMS documentation.
If NULL, this function will generate a "}.outer" file automatically; the user will be prompted to accept the auto-generated *.outer file prior to running eems.
If you don't want to use the automatically generated *.outer file, you can create a custom one here: https://www.keene.edu/campus/maps/tool/}

\item{ask.use.outer}{When outer = NULL, this logical determines if the auto-generated outer coordinates should be plotted, at which point the user must confirm if they want to use the coordinates. Default is TRUE.}

\item{output.dirpath}{Character string with path to directory where EEMS input and output files should be saved. This directory must not exist prior to running this function, but its parent directory must exist.
Files and directories generated by this function:
/output.dirpath/
/output.dirpath/data
/output.dirpath/data/data.coord
/output.dirpath/data/data.diffs
/output.dirpath/data/data.outer
/output.dirpath/mcmc    ### This will contain as many mcmc subdirectories as the value of nchains.
/output.dirpath/params  ### This will contain as many parameter files as the value of nchains.}

\item{n.sites}{The number of sites. Determined automatically if input data is a genind or vcf object/file, but must be supplied if input.data is a path to a diffs matrix.}

\item{ploidy}{Ploidy of individuals. Default is 2.}

\item{nDemes}{Number of demes. Default is 300. This can be a single number or a numeric vector of length nchains with value to use for each chain.}

\item{numMCMCIter}{Number of MCMC iterations (i.e., chain length).Default is 10000000. This can be a single number or a numeric vector of length nchains.}

\item{numBurnIter}{Number of Burnin iterations (i.e., chain length). Default is 1000000. This can be a single number or a numeric vector of length nchains.}

\item{numThinIter}{Number of iterations to ignore before sampling the next MCMC iteration. Default is 9999. This can be a single number or a numeric vector of length nchains.}

\item{nchains}{Number of chains to run. Default is 3.}

\item{setup.only}{If the function should setup data, parameters, and output directories but not run EEMS. Default is TRUE}
}
\value{
Nothing is returned.
}
\description{
runEEMs_snps function

Wrapper function for running EEMS. Optionally generates the '\emph{.outer' file from the '}.coords' file automatically.
}
