% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runeems_snps_setup.R
\name{runeems_snps_setup}
\alias{runeems_snps_setup}
\title{runeems_snps_setup function

Wrapper function for preparing to run EEMS. A bash script is produced that will run EEMS. Optionally generates a '\emph{.outer' file from the '}.coords' file.}
\usage{
runeems_snps_setup(
  x,
  coords,
  save.in,
  outer = NULL,
  exe.path = NULL,
  n.sites = NULL,
  pl = 2,
  nDemes = 300,
  numMCMCIter = 1e+07,
  numBurnIter = 1e+06,
  numThinIter = 9999,
  nchains = 3,
  ...
)
}
\arguments{
\item{x}{An object of class genind or vcfR, or a character string with path to a diffs or VCF file.}

\item{coords}{Character string with path to coordinates file, which has two columns with longitude and latitude coordinates (decimal degree format) of individuals in the genind object. Columns should be space or tab separated.}

\item{save.in}{Character string with path to directory where EEMS input and output files should be saved. This directory must not exist prior to running this function, but its parent directory must exist.
Files and directories generated by this function:
/save.in/
/save.in/data
/save.in/data/data.coord
/save.in/data/data.diffs
/save.in/data/data.outer
/save.in/mcmc    ### This will contain as many mcmc subdirectories as the value of nchains.
/save.in/params  ### This will contain as many parameter files as the value of nchains.}

\item{outer}{Either NULL (the default) or a character string with path to the '\emph{.outer' file, which has two columns with longitude and latitude coordinates (decimal degree format) defining the perimeter of a polygon covering the region to model with EEMS. See EEMS documentation.
If NULL, this function will generate a "}.outer" file automatically; the user will be prompted to accept the auto-generated *.outer file prior to running eems.
If you don't want to use the automatically generated *.outer file, you can create a custom one here: https://www.keene.edu/campus/maps/tool/}

\item{exe.path}{Character string with path to runeems_snps executable, or NULL (the default) if you've defined the EEMS path with the function config_miscwrappers. Use paths_misc.wrappers() to check if this path has been set.}

\item{n.sites}{The number of sites. Determined automatically if 'x' is a VCF file or vcfR or genind object, but must be supplied if 'x' is a path to a diffs matrix.}

\item{pl}{Ploidy. Default is 2.}

\item{nDemes}{Number of demes. Default is 300. This can be a single number or a numeric vector of length nchains with value to use for each chain.}

\item{numMCMCIter}{Number of MCMC iterations (i.e., chain length). Default is 10000000. This can be a single number or a numeric vector of length nchains.}

\item{numBurnIter}{Number of Burnin iterations (i.e., chain length). Default is 1000000. This can be a single number or a numeric vector of length nchains.}

\item{numThinIter}{Number of iterations to ignore before sampling the next MCMC iteration. Default is 9999. This can be a single number or a numeric vector of length nchains.}

\item{nchains}{Number of chains to run. Default is 3.}

\item{...}{Arguments to pass to create.outer. See ?create.outer() for possible arguments.}
}
\value{
Nothing is returned.
}
\description{
runeems_snps_setup function

Wrapper function for preparing to run EEMS. A bash script is produced that will run EEMS. Optionally generates a '\emph{.outer' file from the '}.coords' file.
}
