% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.outer.R
\name{rx2y2}
\alias{rx2y2}
\title{Random point conditioned on distance

Return x and y coordinates of random points conditional on a distance or range of distances from input point.}
\usage{
rx2y2(p1, d, n = 1)
}
\arguments{
\item{p1}{Numerical vector with decimal longitude and latitude of first point.}

\item{d}{Number specifying the distance (in decimal degrees) between the new point and the point with coordinates 'x1y1', or, a numerical vector, in which case the distance from x1y1 will be randomly drawn from values in the range of d.}

\item{n}{Number of points to return. Default 1.}
}
\value{
Numerical vector with longitude and latitude of new point.
}
\description{
Random point conditioned on distance

Return x and y coordinates of random points conditional on a distance or range of distances from input point.
}
