% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPC_adegenet.R
\name{sim.vcf}
\alias{sim.vcf}
\title{Simulate SNPs as VCF}
\usage{
sim.vcf(
  x = NULL,
  save.as = NULL,
  RA.probs = "equal",
  n.ind = NULL,
  n.snps = NULL,
  snp.str = 0,
  ploidy = NULL,
  K = 2,
  ...
)
}
\arguments{
\item{x}{Path to input VCF or a vcfR object. Default NULL. Used to extract probabilities that a site has a particular pair of Reference and Alternative nucleotides. If NULL 'probs' must be supplied.}

\item{save.as}{Null or a character string where the output VCF containing the simulated dataset should be saved.}

\item{RA.probs}{Either one of the character strings "equal" or "empirical", or a numerical vector with length 12 that sums to 1; each value is probability that a SNP has reference allele = i, alternative allele = j, for all [i,j]{A,C,G,T}. If elements of the vector are unnamed, assumed names will be c("AC","AG","AT","CA","CG","CT","GA","GC","GT","TA","TC","TG"), with the first character specifying the reference allele and the second the alternative allele.
The default value of RA.probs is 'equal', which is equivalent to RA.probs=sapply(c("AC","AG","AT","CA","CG","CT","GA","GC","GT","TA","TC","TG"),assign,(1/12)).
If RA.probs="empirical", the frequencies of each Reference:Alternative allele combination is calculated for the input dataset and used for sampling probabilities for the simulated dataset.}

\item{n.ind}{Number of individuals to include in simulated dataset. Default NULL, in which case the number of simulated individuals will match the number of individuals in 'x'.}

\item{n.snps}{Number of SNPs. If n.snp.nonstruc is NULL and x is non-NULL, n.snp.nonstruc will equal half the number of SNPs in the input dataset.}

\item{snp.str}{Fraction of SNPs that are structured \link{WITHIN?} populations. Default 0. Increasing more than 0 seems generate at least twice as many simulated populations as specified with K.}

\item{ploidy}{Number indicating ploidy of individuals. Probably can only be 1 or 2?}

\item{K}{Number of populations (>=2). Default 2.}

\item{...}{Additional arguments passed to the function 'glSim' from 'adegenet' package.}

\item{n.snp.nonstruc}{Number of nonstructured SNPs. If n.snp.nonstruc is NULL and x is non-NULL, n.snp.nonstruc will equal half the number of SNPs in the input dataset.}

\item{n.snp.struc}{Number of structured SNPs. If n.snp.nonstruc is NULL and x is non-NULL, 'n.snp.nonstruc' will equal the number of SNPs in the input dataset minus the value of 'n.snp.nonstruc'. Meaningless if K = 1.}
}
\value{
An object with class vcfR (see 'vcfR' package for details regarding this class)
}
\description{
Create a simulated SNP dataset from either an input VCF file, an object of class vcfR, or by specifying characteristics of the simulated dataset using the function arguments.
Most of the work is done by a call to the function 'glSim' from the 'adegenet' package.
Currently, including missing data in simulated datasets is not possible.
}
