% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp2coords.R
\name{sp2coords}
\alias{sp2coords}
\title{sp2coords function

Function to extract the coordinates matrix from a class Polygon, Polygons, SpatialPolygons, or SpatialPolygonsDataFrame object that containins a single polygon.
Warning: Do not use this function if the inout spatial object contains more than coordinates matrix.}
\usage{
sp2coords(x)
}
\arguments{
\item{x}{sp::Spatial* class object ultimately determined by a single coordinates matrix.}
}
\value{
Numerical matrix of coordinates with longitude and latitude columns, respectively; coordinates must be in decimal degrees format.
}
\description{
sp2coords function

Function to extract the coordinates matrix from a class Polygon, Polygons, SpatialPolygons, or SpatialPolygonsDataFrame object that containins a single polygon.
Warning: Do not use this function if the inout spatial object contains more than coordinates matrix.
}
