% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_miscwrappers.R
\name{sq}
\alias{sq}
\title{sh single quote

Calls the base function shQuote with default options; sq is faster to type.}
\usage{
sq(x)
}
\arguments{
\item{x}{Character string to be quoted}
}
\value{
Character string with leading and trailing "'" character added to input string
}
\description{
sh single quote

Calls the base function shQuote with default options; sq is faster to type.
}
\examples{
library(misc.wrappers)

# Example 1: view current path settings
config_miscwrappers()
# Example 2: set path to program 'runeems_snps' (replace example path with the real path on you machine)
config_miscwrappers(exe.path="/FULL/PATH/TO/runeems_snps")
# Example 3: set path to the fastStructure file 'structure.py'
config_miscwrappers(exe.path="/FULL/PATH/TO/structure.py")
}
