% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{sstacks_cleanup}
\alias{sstacks_cleanup}
\title{cleanup sstacks

Deletes sstacks bash scripts in a directory. Run this after all sstacks jobs are completed.}
\usage{
sstacks_cleanup(dir)
}
\arguments{
\item{dir}{Character string with path to directory containing sstacks bash scripts}
}
\description{
cleanup sstacks

Deletes sstacks bash scripts in a directory. Run this after all sstacks jobs are completed.
}
