% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{sstacks_setup}
\alias{sstacks_setup}
\title{batch setup sstacks

For each set samples (processed with ustacks and cstacks) and a population map, this function creates a bash job file to run sstacks.
The output of this function is a character string vector with paths to bash job files, which can be piped to the function 'rsbatch' to submit the jobs.}
\usage{
sstacks_setup(save.in, popmap = NULL, p = 16, sstacks.path = NULL)
}
\arguments{
\item{save.in}{Character string with directory where output directories and files should be saved. This should be the same as the value used for the 'save.in' argument of ustacks_setup.}

\item{popmap}{Character string to population map. Default NULL, in which case all individuals in 'save.in' are assigned to the same 'population'.}

\item{p}{Number of threads to use for job parallization. Default 16.}

\item{sstacks.path}{Character string with path to the sstacks executable. Default NULL, in which case the path is derived from the default setting, which can be set with config_miscwrappers('<path/to/sstacks>').}
}
\value{
Character string vector with the paths to bash job files.
}
\description{
batch setup sstacks

For each set samples (processed with ustacks and cstacks) and a population map, this function creates a bash job file to run sstacks.
The output of this function is a character string vector with paths to bash job files, which can be piped to the function 'rsbatch' to submit the jobs.
}
