% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{stacks_cleanup}
\alias{stacks_cleanup}
\title{cleanup stacks

Deletes cstacks, sstacks, tsv2bam, gstacks, and populations bash scripts in a directory. Run this after completing the entire STACKS pipeline.}
\usage{
stacks_cleanup(dir)
}
\arguments{
\item{dir}{Character string with path to directory containing populations bash scripts}
}
\description{
cleanup stacks

Deletes cstacks, sstacks, tsv2bam, gstacks, and populations bash scripts in a directory. Run this after completing the entire STACKS pipeline.
}
