% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{summarize_stacks}
\alias{summarize_stacks}
\title{summarize_stacks

For each set samples (processed with populations and with vcf file generated for snps) this function creates a bash job file to calculate basic statistics.}
\usage{
summarize_stacks(dir, save = TRUE, use.popmap = FALSE, popmap = NULL)
}
\arguments{
\item{dir}{Directory containing multiple subdirectories, each of which contains the output files generated by populations (stacks pipeline).}

\item{save}{Logical indicating if the output should be saved. Default TRUE.}

\item{use.popmap}{Logical indicating whether or not to assign individuals to different populations according to a population map. Default FALSE.}

\item{popmap}{Character string to population map. Default NULL, in which case all individuals in 'save.in' are assigned to the same 'population'.}
}
\value{
data frame with basic stats
}
\description{
summarize_stacks

For each set samples (processed with populations and with vcf file generated for snps) this function creates a bash job file to calculate basic statistics.
}
