% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{summarize_VCFs}
\alias{summarize_VCFs}
\title{summarize_VCFs

Same as 'summarize_stacks', except the full paths to input VCFs and an output directory are suppllied as arguments.}
\usage{
summarize_VCFs(VCF.paths, save.in, popmap.path = NULL)
}
\arguments{
\item{VCF.paths}{Character string vector to one or more input VCF files.}

\item{save.in}{Path to directory where the output table should be save. If NULL, the the table is not saved.}

\item{popmap.path}{Character string to population map. If NULL, all individuals in are assigned to the same 'population'.}
}
\value{
data frame with basic stats
}
\description{
summarize_VCFs

Same as 'summarize_stacks', except the full paths to input VCFs and an output directory are suppllied as arguments.
}
