% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{sumnorm.PDF}
\alias{sumnorm.PDF}
\title{PDF sum of normal distributions

Probability distribution for the sum of n normal distributions}
\usage{
sumnorm.PDF(x, m, v)
}
\arguments{
\item{x}{sample quantile}

\item{m}{numerical vector with means of the normal distributions}

\item{v}{numerical vector with variances of the normal distributions}
}
\value{
probability distribution at x for the sum of the normal distributions with means m and variances v
}
\description{
PDF sum of normal distributions

Probability distribution for the sum of n normal distributions
}
