% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{theme_void}
\alias{theme_void}
\title{theme void

mostly blank ggplot theme
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').}
\usage{
theme_void()
}
\value{
ggplot object
}
\description{
theme void

mostly blank ggplot theme
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').
}
