% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{tiles2contours}
\alias{tiles2contours}
\title{tiles2contours

Converts tiles two contours}
\usage{
tiles2contours(tiles, rates, seeds, marks, distm)
}
\arguments{
\item{tiles}{voronoi tiles}

\item{rates}{voronoi rates}

\item{seeds}{Two column matrix with voronoi xseeds and voronoi yseeds}

\item{marks}{'marks' named entry of output list of read_dimns function}

\item{distm}{'dist_metric' named entry of output list of read_dimns function}
}
\value{
countours
}
\description{
tiles2contours

Converts tiles two contours
}
