% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmaphelper.R
\name{transitionModels}
\alias{transitionModels}
\title{Generate transition models

Makes a list of all possible model schemes for transitions between a set of states. The returned object can be passed to the function fitmodels}
\usage{
transitionModels(states, include.asymmetric = TRUE, constraint = NULL)
}
\arguments{
\item{states}{A vector with values representing character states.}

\item{include.asymmetric}{Whether or not to include assymetric rate models in the list of returned models. Default TRUE; (FALSE not yet implemented)}

\item{constraint}{A numeric matrix with values 1 and/or 0, indicating whether or not specific transition types should be allowed or restricted, respectively. The default value of NULL allows all model types in the returned list of models. Providing a constraint matrix filters the set of models to only those that satisfy the constraint.}
}
\value{
A list of transition matrices, for all possible partitions of pairwise transitions into parameters describing transitions.
}
\description{
Generate transition models

Makes a list of all possible model schemes for transitions between a set of states. The returned object can be passed to the function fitmodels
}
