% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{tsv2bam_cleanup}
\alias{tsv2bam_cleanup}
\title{cleanup tsv2bam

Deletes tsv2bam bash scripts in a directory. Run this after all tsv2bam jobs are completed.}
\usage{
tsv2bam_cleanup(dir)
}
\arguments{
\item{dir}{Character string with path to directory containing tsv2bam bash scripts}
}
\description{
cleanup tsv2bam

Deletes tsv2bam bash scripts in a directory. Run this after all tsv2bam jobs are completed.
}
