% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_miscwrappers.R
\name{update_miscwrappers}
\alias{update_miscwrappers}
\title{update miscwrappers

Updates misc.wrappers by calling remotes::install_github function}
\usage{
update_miscwrappers(upgrade = FALSE, ...)
}
\arguments{
\item{upgrade}{Whether or not to update dependencies. Default FALSE.}

\item{...}{Additional arguments to pass to install_github}
}
\description{
update miscwrappers

Updates misc.wrappers by calling remotes::install_github function
}
