% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{ustacks_setup}
\alias{ustacks_setup}
\title{batch setup ustacks

This function creates a bash job file for each combination of input sequence read files (one per individual), values of m and M to run STACKS.
The output of this function is a character string vector with paths to bash job files, which can be piped to the function 'rsbatch' to submit the jobs.}
\usage{
ustacks_setup(ustacks.path, files, save.in, mM, p = 16)
}
\arguments{
\item{ustacks.path}{Character string with path to the ustacks executable.}

\item{files}{Input sequence data files, one for each individual.}

\item{save.in}{Character string with directory where output directories and files should be saved.}

\item{mM}{Two-column data frame or matrix with m and M values in the first and second columns, respectively. Each row indicates a combination of m and M to use for STACKS}

\item{p}{Number of threads to use for job parallization. Default 16.}
}
\value{
Character string vector with the paths to bash job files.
}
\description{
batch setup ustacks

This function creates a bash job file for each combination of input sequence read files (one per individual), values of m and M to run STACKS.
The output of this function is a character string vector with paths to bash job files, which can be piped to the function 'rsbatch' to submit the jobs.
}
