% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fastStructure.R
\name{vcf2fastStructure}
\alias{vcf2fastStructure}
\title{Convert VCF file/object to fastStructure input file (for SNP data)

Converts a VCF file or vcfR object with SNP data into a fastStructure file.}
\usage{
vcf2fastStructure(vcf, IndvNames = TRUE, out = NULL, OtherData = NULL)
}
\arguments{
\item{vcf}{Character string with path to input VCF, or an object of class vcfR.}

\item{IndvNames}{Logical indicating if the first column of the output file should include the names of individuals. Default TRUE.}

\item{out}{Path where output file should be written.}

\item{OtherData}{NULL (the default) or a matrix with up to five (if IndvNames TRUE) or six (if IndvNames FALSE) columns of metadata. If non-NULL, the number of rows must equal number of individuals in the VCF.}
}
\value{
Character string with value of 'out'.
}
\description{
Convert VCF file/object to fastStructure input file (for SNP data)

Converts a VCF file or vcfR object with SNP data into a fastStructure file.
}
\examples{
library(misc.wrappers)
}
