% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions.R
\name{vcf2parts}
\alias{vcf2parts}
\title{vcf2parts

Generate a phylip format partition file from VCF file.}
\usage{
vcf2parts(x, save.as)
}
\arguments{
\item{x}{'vcfR' object (see package::vcfR) or character string with path to VCF file containing nucleotide data.}

\item{save.as}{Path where partition file should be saved.}
}
\value{
Character string with path to partition file.
}
\description{
vcf2parts

Generate a phylip format partition file from VCF file.
}
