% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions.R
\name{vcf2phy}
\alias{vcf2phy}
\title{VCF format to Phylip format

Converts VCF to phylip and saves the result to file. Output can be sequential or interleaved.}
\usage{
vcf2phy(
  x,
  save.as,
  missing = "-",
  sequential = TRUE,
  width = 300,
  padding = 5,
  partition.file = TRUE
)
}
\arguments{
\item{x}{'vcfR' object (see package::vcfR) or character string with path to VCF file containing nucleotide data.}

\item{save.as}{Path where phylip alignment should be saved}

\item{missing}{Character to use for missing data in the output file. Default "-".}

\item{sequential}{Whether or not data should be written on a single line (TRUE) or in multiple blocks (interleaced format) with the number of bases per line controlled by 'width'.}

\item{width}{NULL or a number specifying the number of bases to write per line when output is interleaved.}

\item{padding}{Number of spaces between the longest sample name and the first base.}
}
\value{
The value of 'save.as'.
}
\description{
VCF format to Phylip format

Converts VCF to phylip and saves the result to file. Output can be sequential or interleaved.
}
