% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fastStructure.R
\name{vcf2structure}
\alias{vcf2structure}
\title{Convert VCF file/object to STRUCTURE file (for SNP data)

Converts/writes a VCF file or vcfR object containing SNPs into STRUCTURE format, which is written to a file.
Note: Does not yet handle 'recessive alleles' or phase information (see STRUCTURE manual). This information is optional for STRUCTURE but may be of interest to some users.}
\usage{
vcf2structure(
  vcf,
  IndvNames = TRUE,
  OneRowPerIndv = TRUE,
  MarkerNames = TRUE,
  MissingData = c(-9),
  out,
  InterMarkerDists = FALSE,
  PopData = NULL,
  PopFlag = NULL,
  LocData = NULL,
  Phenotype = NULL,
  OtherData = NULL
)
}
\arguments{
\item{vcf}{Character string with path to input VCF, or an object of class vcfR.}

\item{IndvNames}{Logical indicating if the first column of the STRUCTURE file should contain the names of the individuals, or a character string with names to to use. Default is TRUE, in which case the names will be the names used in the VCF file.}

\item{OneRowPerIndv}{Logical specifying if each genotypes should be written to a single row or multiple rows. Default TRUE, but STRUCTURE works equivalently with either format.}

\item{MarkerNames}{Either a logical (TRUE or FALSE), or a character string with marker names. If TRUE (the default), the marker names are constructed from the CHROM and POS columns of the VCF file: 'CHROM_POS' (e.g. "102_4").}

\item{MissingData}{Number to use for missing-data. Default is -9.}

\item{out}{Path where output structure file should be written. The mainparams file is also written using the same name but with the extension '.params'.}

\item{InterMarkerDists}{Either a logical indicating if intermarker distances should be included from CHROM and POS columns of the VCF, or a vector of integers with inter-marker distances to use. Default is FALSE. A warning is generated if multiple sites per locus are present in the input VCF and 'InterMarkerDists' is FALSE. See STRUCTURE manual for details on supplying inter-marker distances.}

\item{PopData}{Either NULL or an integer vector indicating user-defined population assignments of individuals. The default is NULL (population data not included in output file). If non-NULL, PopData is written in the second column of the output file.}

\item{PopFlag}{Either NULL or a logical vector indicating whether or not STRUCTURE should use the PopData information for the particular individual. Default is NULL, in which case the PopFlag column is not written in the output file. If supplied, PopData must be non-NULL.}

\item{LocData}{NULL (the default) or a vector of integers specifying user-defined sampling locality for each individual.}

\item{Phenotype}{NULL (the default) or a vector of integers specifying the value of a phenotype of interest for each individual.}

\item{OtherData}{NULL (the default) or either a matrix or data frame with as many rows as individuals and columns specifying any other information of interest associated with individuals.}
}
\value{
A list with [\link{1}] Value of 'out', which contains the path to the output file, and [\link{2}] values that should be used for some of the parameters in the mainparams file.
}
\description{
Convert VCF file/object to STRUCTURE file (for SNP data)

Converts/writes a VCF file or vcfR object containing SNPs into STRUCTURE format, which is written to a file.
Note: Does not yet handle 'recessive alleles' or phase information (see STRUCTURE manual). This information is optional for STRUCTURE but may be of interest to some users.
}
