% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPC_adegenet.R
\name{vcf_getSNP}
\alias{vcf_getSNP}
\title{Get Best SNP for each locus from VCF

From a VCF with multiple sites/locus, create a VCF with only the best site/locus.
The best site is the one with the least missing data. To break ties, take the first site among the best sites.
Note: I may update this to not require vcftools.}
\usage{
vcf_getSNP(
  vcftools.path,
  vcf,
  out,
  indv.keep = NULL,
  which.site = "best",
  min.n = 4,
  min.n0 = 2,
  min.n1 = 1
)
}
\arguments{
\item{vcftools.path}{Character string with path to the vcftools executable.}

\item{vcf}{Character string with path to input vcf.}

\item{out}{Character string where to write output vcf.}

\item{indv.keep}{Character string with names of individuals to keep. Default is NULL (all individuals kept).}

\item{which.site}{Character string indicating the method for choosing a site to keep for each locus (or chromosome). Default = "best", which is considered the one with the least missing data, or the first among sites tied for least missing data. Other options are "all.passing", which retains all sites (positions) that pass variation filters (min.n, min.0.n.0, min.1.n), "first" (first site kept at each locus), or "random".}

\item{min.n}{Minimum number of non-missing alleles required to keep a site. Default = 4. If set to "all", then no sites with any missing data are removed (after first filtering individuals if indv.keep is non-NULL).}

\item{min.n0}{Minimum number of individuals required to have at least one copy of the major allele to keep a site. Default = 2.}

\item{min.n1}{Minimum number of individuals required to have at least one copy of the minor allele to keep a site. Default = 1.}
}
\value{
List with [\link{1}] path to vcftools, [\link{2}] dataframe with input and output values for VCF filepaths, number of loci (chromosomes), sites (positions), and individuals (samples).
}
\description{
Get Best SNP for each locus from VCF

From a VCF with multiple sites/locus, create a VCF with only the best site/locus.
The best site is the one with the least missing data. To break ties, take the first site among the best sites.
Note: I may update this to not require vcftools.
}
