% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SNMF.R
\name{vcfR2geno}
\alias{vcfR2geno}
\title{Convert VCF object/file to genotypic matrix

The object returned by vcfR2geno can be used as input in the LEA function snmf
Optionally supply a character string with path where geno object will be saved
This function is used in the function run_SNMFR}
\usage{
vcfR2geno(vcf, out = NULL)
}
\arguments{
\item{vcf}{Character string with path to input VCF file.}

\item{out}{Character string with path where output geno file should be saved. Default is NULL.}
}
\value{
Matrix with genotypes in geno format
}
\description{
Convert VCF object/file to genotypic matrix

The object returned by vcfR2geno can be used as input in the LEA function snmf
Optionally supply a character string with path where geno object will be saved
This function is used in the function run_SNMFR
}
\examples{
library(ade4)
library(adegenet)
library(vcfR)
library(ggplot2)
library(maps)
library(geosphere)
library(tess3r)
library(LEA)
library(rworldmap)
library(JeffWeinell/misc.wrappers)
#source("DAPC_adegenet.R")
#source("runtess.R")
#source("run_SNMF.R")

dev.new(width=10,height=6)
leporinum <- run_SNMF(vcf="Oxyrhabdium-leporinum_BestSNP.vcf",
                      coords="Oxyrhabdium-leporinum_coords.txt",
                      save.as="Oxyrhabdium-leporinum_BestSNP_sNMF_run3.pdf")

modestum  <- run_SNMF(vcf="Oxyrhabdium-modestum_BestSNP.vcf",
                     coords="Oxyrhabdium-modestum_coords.txt",
                     save.as="Oxyrhabdium-modestum_BestSNP_sNMF_run1.pdf")

cfmodestum  <- run_SNMF(vcf="Oxyrhabdium-cf.modestum_BestSNP.vcf",
                     coords="Oxyrhabdium-cfmodestum_coords.txt",
                     save.as="Oxyrhabdium-cfmodestum_BestSNP_sNMF_run1.pdf")

cfmodestum_Luzon     <- run_SNMF(vcf="Oxyrhabdium-cf.modestum_Luzon_BestSNP.vcf",
                        coords="Oxyrhabdium-cfmodestum_Luzon_coords.txt",
                        save.as="Oxyrhabdium-cf.modestum_Luzon_BestSNP_sNMF_run1.pdf")

leporinum_Luzon <- run_SNMF(vcf="Oxyrhabdium-leporinum_Luzon_BestSNP.vcf",
                           coords="Oxyrhabdium-leporinum_Luzon_coords.txt",
                           save.as="Oxyrhabdium-leporinum_Luzon_BestSNP_sNMF_run1.pdf")

bothmodestum <- run_SNMF(vcf="Oxyrhabdium_both-modestum_BestSNP.vcf",
                           coords="Oxyrhabdium_bothmodestum_coords.txt",
                           save.as="Oxyrhabdium_both-modestum_BestSNP_sNMF_run1.pdf")

Oxyrhabdium <- run_SNMF(vcf="Oxyrhabdium_AllSpecies_BestSNP.vcf",
                            coords="Oxyrhabdium_AllSpecies_coords.txt",
                            save.as="Oxyrhabdium_AllSpecies_BestSNP_sNMF_v1.pdf")

Oxyrhabdium_AllSpecies <- run_SNMF(vcf="Oxyrhabdium_AllSpecies_BestSNP.vcf",
                         coords="Oxyrhabdium_AllSpecies_coords.txt",
                         save.as="Oxyrhabdium_AllSpecies_BestSNP_sNMF_run1.pdf")
}
