% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtess.R
\name{vcfR2lfmm}
\alias{vcfR2lfmm}
\title{Reads a VCF object (including those not usually compatible with LEA) and returns a genotypic matrix equivalent to the lfmm-format used by LEA

The object returned by vcfR2lfmm can be used as input in the LEA function tess3.
Optionally supply a character string with path where lfmm object will be saved
This function is used in the function runtess}
\usage{
vcfR2lfmm(vcf, out = NULL)
}
\arguments{
\item{vcf}{Character string with path to input VCF file.}

\item{out}{Character string with path where output lfmm file should be saved. Default is NULL.}
}
\value{
Matrix with genotypes in lfmm format
}
\description{
Reads a VCF object (including those not usually compatible with LEA) and returns a genotypic matrix equivalent to the lfmm-format used by LEA

The object returned by vcfR2lfmm can be used as input in the LEA function tess3.
Optionally supply a character string with path where lfmm object will be saved
This function is used in the function runtess
}
