% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPC_adegenet.R
\name{dapc.plot.arrange}
\alias{dapc.plot.arrange}
\title{Arrange DAPC density plots.

Function to arrange plots of density for each DF or PC and each K}
\usage{
\method{dapc}{plot.arrange}(
  x,
  variable = "DF",
  layout.mat = NULL,
  pos.x.labs = 1,
  pos.y.labs = 2,
  row.labels.left = NULL,
  col.labels.top = NULL,
  row.labels.right = NULL,
  col.labels.bottom = NULL,
  use.diag = NULL,
  pad = 0.1,
  K = NULL,
  outer.text = list(NULL, NULL, NULL, NULL),
  maxMat = c(5, 5)
)
}
\arguments{
\item{x}{A list of ggplots produced by ggscatter.dapc function}

\item{variable}{Either "DF" or "PC"}

\item{layout.mat}{Custom layout matrix to use. Default is NULL, in which case the layout will be generated automatically as a fucnton of the number of plots.}

\item{pos.x.labs}{Where the default labels x axis labels (i.e., column labels) should be drawn. Default = 1 (below table of plots); alternatively 3 (above plot).}

\item{pos.y.labs}{Where the default labels y axis labels (i.e., row labels) should be drawn. Default = 2 (left side); alternatively 3 (right side).}

\item{row.labels.left}{Labels to use the left of the first column of plots. Default NULL (no labels).}

\item{col.labels.top}{Text labels to use above the first row of plots. Default NULL (no labels).}

\item{row.labels.right}{Text labels to to the right of the last column of plots. Default NULL (no labels).}

\item{col.labels.bottom}{Text labels to use below the bottom row of plots. Default NULL (no labels).}

\item{use.diag}{NULL (the default) or numerical vector with values in 1:4, indicating which sides of the gtable (1=bottom, 2=left, 3=top, 4=right) should have labels applied to the diagonal rather than the table margin plots. Ignored if layout.mat is not square. For example, if use.diag=1, then the values of col.labels.bottom will appear below the plots on the diagonal rather than below the plots of the bottom row.}

\item{pad}{Amount of space between plots, in units of line widths (Default 0.1).}

\item{K}{Which set of biplots to use. If NULL (the default), the function will attemp to draw all biplots for all K (max 25 plots).}

\item{outer.text}{A list with length=4, with each entry either NULL (the default) or character string to use as labels below, left, above, and to the right of the arrangement of plots.}

\item{maxMat}{Numerical vector with length 2 that specifies the maximum number of rows and columns of plots, respectively, in the output gtable. Default = c(7,8).}
}
\value{
A gtable object
}
\description{
dev.new(height=10,width=6)
Ahaetulla_Luzon <- run_DAPC(vcf="Ahaetulla-prasina_Luzon_1snpPerLocus.vcf",
probs.out="Ahaetulla-prasina_Luzon_BestSNP_DAPC_BestK_Membership_run2.txt",
save.out="Ahaetulla-prasina_Luzon_DAPC_v2.pdf")
Calamaria_allpops <- run_DAPC(vcf="Calamaria-gervaisii_snps.vcf",
probs.out="Calamaria-gervaisii_allpops_BestSNP_DAPC_BestK_Membership_run2.txt",
save.as="Calamaria-gervaisii_allpops_DAPC.pdf")
dev.new(height=6,width=10)
cfmodestum_Luzon <- run_DAPC(vcf="Oxyrhabdium-cf.modestum_Luzon_BestSNP.vcf",
probs.out="Oxyrhabdium-cf.modestum_Luzon_BestSNP_DAPC_BestK_Membership_run2.txt",
coords="tess3r/Oxyrhabdium-cfmodestum_Luzon_coords.txt",
save.as="Oxyrhabdium-cfmodestum_Luzon_BestSNP_DAPC_run2.pdf")
leporinum_Luzon <- run_DAPC(vcf="Oxyrhabdium-leporinum_Luzon_BestSNP.vcf".
probs.out="Oxyrhabdium-leporinum_Luzon_BestSNP_DAPC_BestK_Membership_run2.txt",
coords="Oxyrhabdium-leporinum_Luzon_coords.txt",
save.as="Oxyrhabdium-leporinum_Luzon_BestSNP_DAPC.pdf")
leporinum  <- run_DAPC(vcf="Oxyrhabdium-leporinum_BestSNP.vcf",
probs.out="Oxyrhabdium-leporinum_BestSNP_DAPC_BestK_Membership_run2.txt",
coords="Oxyrhabdium-leporinum_coords.txt",
save.as="Oxyrhabdium-leporinum_BestSNP_DAPC_run2.pdf")
modestum   <- run_DAPC(vcf="Oxyrhabdium-modestum_BestSNP.vcf",
probs.out="Oxyrhabdium-modestum_BestSNP_DAPC_BestK_Membership.txt_run2",
coords="Oxyrhabdium-modestum_coords.txt",
save.as="Oxyrhabdium-modestum_BestSNP_DAPC_run2.pdf")
cfmodestum <- run_DAPC(vcf="Oxyrhabdium-cf.modestum_BestSNP.vcf",
probs.out="Oxyrhabdium-cf.modestum_BestSNP_DAPC_BestK_Membership_run2.txt",
coords="Oxyrhabdium-cfmodestum_coords.txt",
save.as="Oxyrhabdium-cf.modestum_BestSNP_DAPC_run2.pdf")
Oxyrhabdium <- run_DAPC(vcf="Oxyrhabdium_AllSpecies_BestSNP.vcf",
probs.out="Oxyrhabdium_AllSpecies_BestSNP_DAPC_BestK_Membership_run2.txt",
coords="Oxyrhabdium_AllSpecies_coords.txt",
save.as="Oxyrhabdium_AllSpecies_BestSNP_DAPC_v3.pdf")
}
