% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{filled_contour_graph}
\alias{filled_contour_graph}
\title{filled contour graph

Generates a filled countour graph ggplot object
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').}
\usage{
filled_contour_graph(p, dimns, plot_params)
}
\arguments{
\item{p}{}

\item{dimns}{output of read_dimns function}

\item{plot_params}{List of plotting parameters}
}
\value{
ggplot object
}
\description{
filled contour graph

Generates a filled countour graph ggplot object
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').
}
