% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{filled_eems_contour}
\alias{filled_eems_contour}
\title{filled eems contour

generates a ggplot object with filled contours}
\usage{
filled_eems_contour(dimns, zmean, plot_params, is_mrates)
}
\arguments{
\item{dimns}{output of read_dimns function}

\item{zmean}{}

\item{plot_params}{List of plotting parameters}

\item{is_mrates}{Logical indicating if the plots are mrates, else qrates}
}
\value{
ggplot object
}
\description{
filled eems contour

generates a ggplot object with filled contours
}
