% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fastStructure.R
\name{find.exe.path}
\alias{find.exe.path}
\title{Find path to program

This function searches the system path and the settings file of the misc.wrappers program for a usable path to a program with a specified name.}
\usage{
find.exe.path(program)
}
\arguments{
\item{program}{Character string with name of program to find}
}
\value{
Character string with path to the program's executable file, or 1 if path not found.
}
\description{
Find path to program

This function searches the system path and the settings file of the misc.wrappers program for a usable path to a program with a specified name.
}
\examples{
library(misc.wrappers)
run_fastStructure
leporinum <- run_fastStructure(vcf="Oxyrhabdium-leporinum_BestSNP.vcf",
          coords="Oxyrhabdium-leporinum_coords.txt",
          kmax=1:15,
          save.as="Oxyrhabdium-leporinum_BestSNP_tess3r_run3.pdf",
          python.path = "/panfs/pfs.local/software/7/install/anaconda/4.7/envs/py27/bin/python",
          fastStructure.path = "/panfs/pfs.local/work/bi/bin/fastStructure-master/")

}
