% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPC_adegenet.R
\name{hex2dec}
\alias{hex2dec}
\title{Hex to xyz colors

Hex to xyz color coordinates}
\usage{
hex2dec(hex)
}
\arguments{
\item{hex}{Character string with hex color code}
}
\value{
xyz color coordinates
}
\description{
da.biPlot; pca.biPlot
x=da.biPlot
}
