% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{load_required_packages}
\alias{load_required_packages}
\title{load required packages

Load packages
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').}
\usage{
load_required_packages(packages)
}
\arguments{
\item{packages}{Character string with names of packages to load.}
}
\value{
message stating packages loaded
}
\description{
load required packages

Load packages
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').
}
