% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{plot_eems}
\alias{plot_eems}
\title{plot_eems function

Invokes the reemsplots2 function make_eems_plots, with additional functionality such as adding island borders and points showing individual coordinates..}
\usage{
plot_eems(
  mcmcdir,
  save.in = NULL,
  plot.coords = T,
  plot.geography = T,
  mask.oceans = T,
  include.out = c("pdf", "raster")
)
}
\arguments{
\item{mcmcdir}{Character string to directory of a particular mcmc chain}

\item{save.in}{Path to directory where output files should be saved. Default is NULL, in which case the value of 'mcmcdir' is used.}

\item{plot.coords}{Whether or not to overlay the sample coordinates on EEMS maps. Default TRUE.}

\item{plot.geography}{Whether or not to overlay country borders on EEMS maps. Default TRUE.}

\item{mask.oceans}{Whether or not mask oceans on EEMS maps. Default TRUE.}

\item{include.out}{Character string vector indicating the type of output files to generate.}
}
\value{
List of ggplots
}
\description{
plot_eems function

Invokes the reemsplots2 function make_eems_plots, with additional functionality such as adding island borders and points showing individual coordinates..
}
