% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.outer.R
\name{points.on.land}
\alias{points.on.land}
\title{Return coordinates over land

Returns the set of input coordinates that intersect with land. Land is considered to incude all areas covered by a feature in the rnaturalearth::ne_countries 10 meter resolution dataset. Input coordinates are are assumed to be WGS84 (EPSG:4326)}
\usage{
\method{points}{on.land}(x, return.as = "data.frame")
}
\arguments{
\item{x}{Matrix of data frame with coordinates as longitude and latitude columns.}

\item{return.as}{Character string with class to use for object returned. Default "data.frame". Can also be "matrix" or "SP" (SpatialPoints).}
}
\value{
An object with class equal to the value of 'return.as', which contains the subset of input points that occur on land, or more specifically, not in an ocean.
}
\description{
Return coordinates over land

Returns the set of input coordinates that intersect with land. Land is considered to incude all areas covered by a feature in the rnaturalearth::ne_countries 10 meter resolution dataset. Input coordinates are are assumed to be WGS84 (EPSG:4326)
}
\examples{
### Load example dataset with coordinates in a two-column matrix (first column longitude).
LonLat <- misc.wrappers::grp1.lonlat
### Get the subset of points that occur on land.
coords.df <- points.on.land(x=LonLat)
### Create the outer set of points and generate a figure with points on map.
createouter_example_method1 <- create.outer(coords= coords.df,method=1,output.path="createouter_exampleOutput_method1_coords.txt",plot.output.path="createouter_exampleOutput_method1_coords.pdf")
}
