% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPC_adegenet.R
\name{pset}
\alias{pset}
\title{Powerset

Function to return the powerset for a numeric or character vector of set elements.}
\usage{
pset(x)
}
\arguments{
\item{x}{Numerical or character vector}
}
\value{
List with powerset of x
}
\description{
Powerset

Function to return the powerset for a numeric or character vector of set elements.
}
