% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.outer.R
\name{rcoords}
\alias{rcoords}
\title{Randomly sample points within a random-centered circle.

Sample points in an area with specified radius and center sampled within an extent, with the optional condition that returned points must be over land.}
\usage{
rcoords(
  r,
  size,
  return.as = "data.frame",
  limits = c(-180, 180, -90, 90),
  over.land = TRUE
)
}
\arguments{
\item{r}{radius of area to sample within, in decimal degrees.}

\item{size}{Number of points to return.}

\item{return.as}{Character string with class to use for object returned. Default "data.frame". Can also be "matrix" or "SP" (SpatialPoints).}

\item{limits}{Numerical vector of length four defining the limits for the sampling area's center, with longitude (minimum), longitude (maximum), latitude (minimum), latitude (maximum). Default is c(-180,180,-90,90). Points other than the center of the sampling area can fall outside of this limit.}

\item{over.land}{Logical indicating whether or not all points in the returned data frame should occur over land. Default TRUE.}
}
\value{
An object with class equal to the value of 'return.as', which contains the subset of input points that occur on land, or more specifically, not in an ocean.
}
\description{
Randomly sample points within a random-centered circle.

Sample points in an area with specified radius and center sampled within an extent, with the optional condition that returned points must be over land.
}
