% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.outer.R
\name{rcoords}
\alias{rcoords}
\title{Sample points in a random-centered area or tethered group of areas.

This function provides a way to randomly generate one or more groups of points, such that groups are non-overlapping, and options for controlling proximity, group sizes, and optional requirement that points occur over land.
At present areas are circles projected onto a euclidean plane.}
\usage{
rcoords(
  regionsize,
  samplesize,
  n.grp = 1,
  grp.n.weights = rep(1, n.grp),
  grp.area.weights = rep(1, n.grp),
  wnd = c(-180, 180, -90, 90),
  over.land = TRUE,
  interactions = c(0, 1),
  expf = 8,
  show.plot = FALSE,
  return.as = "data.frame"
)
}
\arguments{
\item{regionsize}{Number specifying the approximate longitudinal diameter spanning all sampling regions. I have not verified the accuracy, but larger values generally yield larger extent across all samples.}

\item{samplesize}{Total number of points to return, across all groups.}

\item{n.grp}{Number of spatial sampling groups.}

\item{grp.n.weights}{Probability weights for determing sample sizes of groups. Default weights are uniformly distributed.}

\item{grp.area.weights}{Probability weights for determing group region sizes. Default weights are uniformly distributed.}

\item{wnd}{Numerical vector specifying longitudinal and latitude limits for possible sampling. Default is to include all of Earth c(-180,180,-90,90).}

\item{over.land}{Logical indicating whether or not all returned points must occur over land. Default TRUE. Note that this condition is only applied to samples returned as output. Therefore, 'samplesize' is really a 'sample until' rule. This is faster than performing clipping operations on proposed sample polygons to conform to geography.}

\item{interactions}{Numeric vector with the maximum and minimum amount of overlap between pairs of groups, calculated as (intersect area)/(mean group areas non-intersected). The default c(0,1) allows for all possible scenarios. Examples: c(0,0) specifies that groups must be allopatric; c(1,1) requires complete overlap of groups, which is not realistic given the stochasticity determining region sizes; c(0.5,1) requires that at least half-overlaps between groups; c(0.2,0.25) specifies a small contact zone.}

\item{expf}{Number that affects dispersion relative to regionsize. Higher numbers increase dispersion. Default 8.}

\item{show.plot}{Whether or not the points should be plotted on a low-resolution land map. The map is used is the rnaturalearth countries map, 110 meter resolution.}

\item{return.as}{Character string with class to use for object returned. Default "data.frame". Can also be "matrix" or "SP" (SpatialPoints).}

\item{r}{Number specifying the radius of group areas, in decimal degrees, or a numerical vector with min and max values of uniform sampling distribution from which radii of areas will be drawn.}

\item{d.grp}{Number controlling the distance between the centers of a pair of areas, as a function of the pair's radii. Default 1, which would allow sample regions to nearly coincide. Future option may allow for a pairwise distance matrix.}
}
\value{
An object with class equal to the value of 'return.as' and containing the set of points that meet the specified sampling requirements. If return.as='matrix' or 'data.frame', the columns are 'X' (for longitude), 'Y' (for latitude), and 'group' (all 1 if 'n.grp'=1).
}
\description{
Sample points in a random-centered area or tethered group of areas.

This function provides a way to randomly generate one or more groups of points, such that groups are non-overlapping, and options for controlling proximity, group sizes, and optional requirement that points occur over land.
At present areas are circles projected onto a euclidean plane.
}
