% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.outer.R
\name{rcoords}
\alias{rcoords}
\title{Sample points in a random-centered area or tethered group of areas.

This function provides a way to randomly generate one or more groups of points, such that groups are non-overlapping, and options for controlling proximity, group sizes, and optional requirement that points occur over land.
At present areas are circles projected onto a euclidean plane.}
\usage{
rcoords(
  r,
  size,
  return.as = "data.frame",
  limits = c(-180, 180, -90, 90),
  over.land = TRUE,
  n.grp = 1,
  interactions = NULL,
  d.grp = 1,
  show.plot = FALSE
)
}
\arguments{
\item{r}{Number specifying the radius of group areas, in decimal degrees, or a numerical vector with min and max values of uniform sampling distribution from which radii of areas will be drawn.}

\item{size}{Number of points to return, or a numerical vector with number of points to return from each adjacent polygon if n.adj=1.}

\item{return.as}{Character string with class to use for object returned. Default "data.frame". Can also be "matrix" or "SP" (SpatialPoints).}

\item{limits}{Numerical vector of length four defining the limits for the sampling area's center, with longitude (minimum), longitude (maximum), latitude (minimum), latitude (maximum). Default is c(-180,180,-90,90). Points other than the center of the sampling area can fall outside of this limit.}

\item{over.land}{Logical indicating whether or not all points in the returned data frame should occur over land. Default TRUE. Note that this condition is only applied to the samples themselves and not the sampling area.}

\item{n.grp}{Number of spatial sampling groups.}

\item{interactions}{Numeric vector with the maximum and minimum amount of overlap required between adjacent groups, calculated as (intersect area)/(sum area) for each pair of groups; or NULL, the default, in which case the all types of group interactions are possible.}

\item{d.grp}{Number controlling the distance between the centers of a pair of areas, as a function of the pair's radii. Default 1, which would allow sample regions to nearly coincide. Future option may allow for a pairwise distance matrix.}

\item{show.plot}{Whether or not the points should be plotted on a low-resolution land map. The map is used is the rnaturalearth countries map, 110 meter resolution.}
}
\value{
An object with class equal to the value of 'return.as' and containing the set of points that meet the specified sampling requirements. If return.as='matrix' or 'data.frame', the columns are 'X' (for longitude), 'Y' (for latitude), and 'group' (all 1 if 'n.grp'=1).
}
\description{
Sample points in a random-centered area or tethered group of areas.

This function provides a way to randomly generate one or more groups of points, such that groups are non-overlapping, and options for controlling proximity, group sizes, and optional requirement that points occur over land.
At present areas are circles projected onto a euclidean plane.
}
