% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eems.R
\name{read_matrix}
\alias{read_matrix}
\title{read matrix

Internal function used for making EEMS plots. Reads eems output file as a matrix.
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').}
\usage{
read_matrix(file, ncol = 2)
}
\arguments{
\item{file}{Path to file to read}

\item{ncol}{Number of columns in the file. Default 2.}
}
\value{
matrix with the file contents
}
\description{
read matrix

Internal function used for making EEMS plots. Reads eems output file as a matrix.
This function is from the R package 'reemsplots2' by Desislava Petkova (github repository 'dipetkov/reemsplots2').
}
