% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SNMF.R
\name{run_SNMF}
\alias{run_SNMF}
\title{Wrapper for LEA snmf

Runs LEA function snmf function from VCF input file and optionally interpolates cluster assignments onto a map (if 'coords' argument is non-NULL).}
\usage{
run_SNMF(
  vcf,
  coords = NULL,
  kmax = 40,
  reps = 100,
  entropy = TRUE,
  project = "new",
  iter = 500,
  save.as = NULL
)
}
\arguments{
\item{vcf}{Path to input VCF file with SNP data}

\item{coords}{Either a character string with path to file containing coordinates (longitude in first column, latitude in second column), or matrix object with longitude and latitude columns.}

\item{kmax}{Number indicating the maximum number of clusters to evaluate. Default is 40, which is converted using kmax = min(kmax, number of individuals-1)}

\item{reps}{Number of repititions. Default 100.}

\item{entropy}{Default TRUE}

\item{project}{Defualt 'new'}

\item{iter}{Default 500}

\item{save.as}{Character string with where to save the output PDF with plots of results. Default is NULL. **Important! This argument is ignored in some environments. Instead, use dev.new(file="Where/To/Save/Output.pdf",height=6,width=10,noRStudioGD=TRUE) before using run_SNMF. Then dev.off().}

\item{CPU}{Defualt 2}
}
\value{
List of plots
}
\description{
Wrapper for LEA snmf

Runs LEA function snmf function from VCF input file and optionally interpolates cluster assignments onto a map (if 'coords' argument is non-NULL).
}
