% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtess.R
\name{runtess}
\alias{runtess}
\title{Run tess3r and plot results

Pipeline with wrappers for several functions from the tess3r package.
Data can be supplied as a VCF file and results are plotted in multiple useful ways.}
\usage{
runtess(
  x,
  format = "VCF",
  coords,
  samplenames = NULL,
  kmax = 10,
  reps = 30,
  save.in = NULL,
  mask = 0.05,
  max.iteration = 500,
  include.out = c(".pdf", ".Qlog", ".entropyLog"),
  overwrite = FALSE
)
}
\arguments{
\item{x}{'vcfR' object (see package::vcfR) or character string with path to SNPs dataset fornatted to match the 'format' argument. Currently only VCF can be used, but others may be added soon.}

\item{format}{Character string indicating the format of the data. Currently only "VCF" or "fastStructure" allowed. Other types may be added. Ignored if x is a vcfR object.}

\item{coords}{Either a character string with path to file containing coordinates (longitude in first column, latitude in second column), or matrix object with longitude and latitude columns.}

\item{samplenames}{NULL or a character string vector with names of samples (in same order) as data in x and coords. If NULL (the default), sample names are extracted from x.}

\item{kmax}{Numerical vector with set of values to use for K. Default 40.}

\item{reps}{Number of repititions. Default 100.}

\item{save.in}{Where to save output files. Default is NULL.}

\item{mask}{Proportion of input data to mask during each replicate when tess3 function is called. Default 0.05.}

\item{max.iteration}{Max iterations. Default 500.}

\item{include.out}{Character vector indicating which type of files should be included as output. Default is c(".pdf",".Qlog", and ".entropyLog").}

\item{overwrite}{Logical indicating whether or not to allow new output files to overwrite existing ones. Default FALSE.}
}
\value{
List of plots
}
\description{
Run tess3r and plot results

Pipeline with wrappers for several functions from the tess3r package.
Data can be supplied as a VCF file and results are plotted in multiple useful ways.
}
