% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacks.R
\name{ustacks_cleanup}
\alias{ustacks_cleanup}
\title{cleanup ustacks

Deletes ustacks bash scripts in a directory. Run this after all ustacks jobs are completed.}
\usage{
ustacks_cleanup(dir)
}
\arguments{
\item{dir}{Character string with path to directory containing ustacks bash scripts}
}
\description{
cleanup ustacks

Deletes ustacks bash scripts in a directory. Run this after all ustacks jobs are completed.
}
