% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SNMF.R
\name{vcfR2geno}
\alias{vcfR2geno}
\title{Convert VCF object/file to genotypic matrix

The object returned by vcfR2geno can be used as input in the LEA function snmf
Optionally supply a character string with path where geno object will be saved
This function is used in the function run_sNMFR}
\usage{
vcfR2geno(vcf, out = NULL)
}
\arguments{
\item{vcf}{Character string with path to input VCF file.}

\item{out}{Character string with path where output geno file should be saved. Default is NULL.}
}
\value{
Matrix with genotypes in geno format
}
\description{
Convert VCF object/file to genotypic matrix

The object returned by vcfR2geno can be used as input in the LEA function snmf
Optionally supply a character string with path where geno object will be saved
This function is used in the function run_sNMFR
}
\examples{
library(misc.wrappers)
# Define path to input VCF file containing similated data for 500 SNPs from 50 individuals in three populations.
example_vcf_path <- file.path(system.file("extdata", package = "misc.wrappers"),"example_simulated_K4.vcf.gz")
# Perform sNMF analyses on the simulated dataset for for K=2–10 and 30 replicates.
test_K4 <- run_sNMF(x=example_vcf_path, kmax=10, reps=30, save.as="sNMF_example.pdf")
}
